% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{accountPreembleMethod2}
\alias{accountPreembleMethod2}
\title{This function applies the account preemble method 2.}
\usage{
accountPreembleMethod2(accounts, sysAccs, custodian, session)
}
\arguments{
\item{accounts}{The accounts list}

\item{sysAccs}{A data frame with the system accounts.}

\item{custodian}{The custodian id.}

\item{session}{The rdecaf session.}
}
\value{
Returns the side by side comparison of decaf position and provider position.
}
\description{
This is a description.
}
