% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdings.R
\name{addTagsAsColumns}
\alias{addTagsAsColumns}
\title{TODO:}
\usage{
addTagsAsColumns(holdings, resources, addTagsBy)
}
\arguments{
\item{holdings}{TODO}

\item{resources}{TODO}

\item{addTagsBy}{TODO}
}
\value{
A data-frame with the holdings
}
\description{
This is the description
}
