% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{advancedStatsHistory}
\alias{advancedStatsHistory}
\title{Wrappper that runs getPerformance historically for portfolio and benchmark when available}
\usage{
advancedStatsHistory(portfolio, date, session, rfs = NULL)
}
\arguments{
\item{portfolio}{the portfolio ID.}

\item{date}{the asofdate.}

\item{session}{The rdecaf session.}

\item{rfs}{the named list indicating what to query for the risk free rate.}
}
\value{
A list with return statistics.
}
\description{
This is the description
}
