% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfra.R
\name{aggEstimates}
\alias{aggEstimates}
\title{CFRA-pocketbase dependent function to unlist the estimates CFRA endpoint. Relevant for cfraDetails function ONLY.}
\usage{
aggEstimates(lst)
}
\arguments{
\item{lst}{list to flatten. Should be the income or earnings from CFRA queryCfra(cc,path)[["result"]], path e.g. equity/research/stars-full/ticker/XXX/exchange/YYY}
}
\value{
data frame of the flattened list.
}
\description{
This is the description
}
