% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregateAndXTSify}
\alias{aggregateAndXTSify}
\title{A function aggregates a value column by a second column and transforms it to xts class.}
\usage{
aggregateAndXTSify(
  x,
  aggColumn = "qtymain",
  aggBy = "commitment",
  fallbackDate = Sys.Date()
)
}
\arguments{
\item{x}{A data frame with the columns names as ascribed to aggColumn and aggBy.}

\item{aggColumn}{The name of the column to be aggregated.}

\item{aggBy}{The name of the column to be aggregated by.}

\item{fallbackDate}{If data frame has NROW==0, we need a fallback date.}
}
\value{
A time-series object of class xts.
}
\description{
This is the description
}
