% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{alert_latest_trades_clever}
\alias{alert_latest_trades_clever}
\title{A function to send latest trades alert.}
\usage{
alert_latest_trades_clever(
  session,
  email_params,
  excl_type = "CCY",
  hours = 24,
  gte = "10:01:00",
  lte = "10:19:00",
  tz = "UTC",
  ...
)
}
\arguments{
\item{session}{The rdecaf session.}

\item{email_params}{The parameters for the email dispatch.}

\item{excl_type}{The resource type to be excluded.}

\item{hours}{The number of hours to look back.}

\item{gte}{Greater than or equal to this time (HH:MM:SS) to run this alert.}

\item{lte}{Less than or equal to this time (HH:MM:SS) to run this alert.}

\item{tz}{The time-zone for gte and lte.}

\item{...}{Andy additional parameters.}
}
\value{
NULL. Email with the alert will be sent.
}
\description{
This is the description
}
