% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{alertBondCoupon}
\alias{alertBondCoupon}
\title{A function to send latest trades alert.}
\usage{
alertBondCoupon(
  session,
  resources,
  days,
  emailParams,
  greeting,
  deployment,
  url,
  gte,
  lte,
  tz = "UTC",
  adhoc
)
}
\arguments{
\item{session}{The rdecaf session.}

\item{resources}{The data-frame of the resources in the decaf instance.}

\item{days}{The number of days until the expiry.}

\item{emailParams}{The parameters for the email dispatch.}

\item{greeting}{The greetings string.}

\item{deployment}{The name of the deployment / client.}

\item{url}{The url of the deployment.}

\item{gte}{Greater than or equal to this time (HH:MM:SS) to run this alert.}

\item{lte}{Less than or equal to this time (HH:MM:SS) to run this alert.}

\item{tz}{The time-zone for gte and lte.}

\item{adhoc}{If TRUE, email gets send only if there are expiring instruments.}
}
\value{
NULL. Email with the alert will be sent.
}
\description{
This is the description
}
