% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{alertDecafData}
\alias{alertDecafData}
\title{A function to produce any email alert, given a respective function.}
\usage{
alertDecafData(
  session,
  emailParams,
  window = c("00:00:00", "23:59:59", "UTC"),
  send = Sys.Date(),
  date = NULL,
  dataFunc,
  dataParams,
  ...
)
}
\arguments{
\item{session}{The rdecaf session.}

\item{emailParams}{List of the parameters for the email dispatch. Should include greeting, subject, etc.}

\item{window}{String vector of the time window in which the alert to be sent, e.g. c(tz,gtw,lte). Defaults to 24H.}

\item{send}{date value of the date representing the as of date. Defaults to present.}

\item{date}{date value of the date in which to send. Defaults to NULL for all.}

\item{dataFunc}{string of the function name to run to return the data. Should be a function that generates a list with two or less elements (summary and details).}

\item{dataParams}{list of the parameters to supply the function.}

\item{...}{Any additional arguments.}
}
\value{
Sends email with the alert and returns a message.
}
\description{
This is the description
}
