% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{alertEmail}
\alias{alertEmail}
\title{A function that sends out a generic DECAF email alert given parameters.}
\usage{
alertEmail(
  session,
  data = NULL,
  emailParams,
  tmPath,
  dtPath = "details.csv",
  snd = Sys.Date(),
  wdw = c("00:00:00", "23:59:00"),
  dte = NULL,
  tz = "Asia/Singapore",
  ...
)
}
\arguments{
\item{session}{The rdecaf session.}

\item{data}{a list containing 2 elements (summary or #1 will be displayed in email body, details or #2 will be in attachments). Default NULL for just a message.}

\item{emailParams}{The parameters for the email dispatch.}

\item{tmPath}{the location of the template used to stylize the email.}

\item{dtPath}{the path containing the details/attachment data. Defaults to details.csv in the wd.}

\item{wdw}{the time period in which to run the function. Defaults to 24H given a vector of c(gte,lte).}

\item{dte}{the weekday name of the day in which to run this alert. Defaults to NULL for all days.}

\item{tz}{The time-zone for gte and lte. Defaults to Asia/Singapore.}

\item{...}{Any additional arguments.}
}
\value{
Sends email with the alert.
}
\description{
This is the description
}
