% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{alertMissingMonthEndPx}
\alias{alertMissingMonthEndPx}
\title{A function to .}
\usage{
alertMissingMonthEndPx(session, resources, emailParams,
  greeting = "Dear Team,", deployment, url, gte = "00:00:00",
  lte = "23:59:00", tz = "UTC")
}
\arguments{
\item{session}{The rdecaf session.}

\item{resources}{The data-frame of the resources in the decaf instance.}

\item{emailParams}{The parameters for the email dispatch.}

\item{greeting}{The greetings string.}

\item{deployment}{The name of the deployment / client.}

\item{url}{The url of the deployment.}

\item{gte}{Greater than or equal to this time (HH:MM:SS) to run this alert.}

\item{lte}{Less than or equal to this time (HH:MM:SS) to run this alert.}

\item{tz}{The time-zone for gte and lte.}
}
\value{
NULL. Email with the alert will be sent.
}
\description{
This is the description
}
