% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{alertMissingResourceFieldV2}
\alias{alertMissingResourceFieldV2}
\title{A function to report a missing or invalid field in resources.}
\usage{
alertMissingResourceFieldV2(
  session,
  resources = NULL,
  emailParams,
  greeting = "Dear Team,",
  deployment,
  url,
  gte = "00:00:00",
  lte = "23:59:59",
  tz = "UTC",
  stockOnly = TRUE,
  emailList = NULL,
  excludeType = "CCY",
  invalidVals = NULL,
  fieldName,
  ...
)
}
\arguments{
\item{session}{The rdecaf session.}

\item{resources}{The data-frame of the resources in the decaf instance. Default NULL.}

\item{emailParams}{The parameters for the email dispatch.}

\item{greeting}{The greetings string.}

\item{deployment}{The name of the deployment / client.}

\item{url}{The url of the deployment.}

\item{gte}{Greater than or equal to this time (HH:MM:SS) to run this alert.}

\item{lte}{Less than or equal to this time (HH:MM:SS) to run this alert.}

\item{tz}{The time-zone for gte and lte.}

\item{stockOnly}{Shall we check only open positions? Default: TRUE}

\item{emailList}{Optionally, a vector with email addressess to overwrite emailParams.}

\item{excludeType}{The ctypes of resources to be excluded. A vector. Default is 'CCY'.}

\item{invalidVals}{A vector with strings which do not qualify as valid.}

\item{fieldName}{The name of the field to check.}

\item{...}{Any additional arguments.}
}
\value{
Sends email with the alert and returns a message.
}
\description{
This is the description
}
