% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{alertNAVComparison}
\alias{alertNAVComparison}
\title{A function to send the NAV comparison alert.}
\usage{
alertNAVComparison(
  sSession,
  tSession,
  accounts,
  type,
  resources,
  days = 1,
  ccy = "USD",
  sDeployment,
  tDeployment,
  emailParams,
  gte = "00:00:00",
  lte = "23:59:00",
  tz = "Asia/Singapore",
  emailByTeam = FALSE
)
}
\arguments{
\item{sSession}{The source session.}

\item{tSession}{The target session.}

\item{accounts}{The account mapping list.}

\item{type}{The type of container. Either 'accounts' or 'portfolios'.}

\item{resources}{The resources data frame.}

\item{days}{The lookback days for the consolidations.}

\item{ccy}{The currency to be valued in.}

\item{sDeployment}{The source deployment name.}

\item{tDeployment}{The target deployment name.}

\item{emailParams}{The email parameters.}

\item{gte}{Greater than or equal to time.}

\item{lte}{Less than or equal to time.}

\item{tz}{The time zone.}

\item{emailByTeam}{Email segregated by team.}
}
\value{
NULL. Email with the alert will be sent.
}
\description{
This is the description
}
