% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{alertsComplianceBreachInspector}
\alias{alertsComplianceBreachInspector}
\title{A function to push a payload to a decaf instance.}
\usage{
alertsComplianceBreachInspector(
  session,
  resources,
  emailParams,
  occurThreshold = 2,
  greeting,
  deployment,
  url,
  gte,
  lte,
  tz
)
}
\arguments{
\item{session}{The rdecaf session.}

\item{resources}{The data-frame of the resources in the decaf instance.}

\item{emailParams}{The parameters for the email dispatch.}

\item{occurThreshold}{For common instrument culprits, what is the occurance threshold? Default = 2.}

\item{greeting}{The greetings string.}

\item{deployment}{The name of the deployment / client.}

\item{url}{The url of the deployment.}

\item{gte}{Greater than or equal to this time (HH:MM:SS) to run this alert.}

\item{lte}{Less than or equal to this time (HH:MM:SS) to run this alert.}

\item{tz}{The time-zone for gte and lte.}
}
\value{
NULL. Email with the alert will be sent.
}
\description{
This is the description
}
