% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansbacherCommons.R
\name{ansbacherCLIV001Columns}
\alias{ansbacherCLIV001Columns}
\title{A function to return the column mappers}
\usage{
ansbacherCLIV001Columns()
}
\value{
Returns a list with the column mappers.
}
\description{
This is the description
}
