% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansbacherCommons.R
\name{ansbacherColumnMappers}
\alias{ansbacherColumnMappers}
\title{A function to return the column mappers}
\usage{
ansbacherColumnMappers()
}
\value{
Returns a list with the column mappers.
}
\description{
This is the description
}
