% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansbacherCommons.R
\name{ansbacherCreateOtherResource}
\alias{ansbacherCreateOtherResource}
\title{TODO:}
\usage{
ansbacherCreateOtherResource(data, session)
}
\arguments{
\item{data}{TODO}

\item{session}{TODO}
}
\value{
TODO
}
\description{
This is the description
}
