% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansbacherCommons.R
\name{ansbacherITGI002Columns}
\alias{ansbacherITGI002Columns}
\title{A function to return the column mappers}
\usage{
ansbacherITGI002Columns()
}
\value{
Returns a list with the column mappers.
}
\description{
This is the description
}
