% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansbacherCommons.R
\name{ansbacherParser}
\alias{ansbacherParser}
\title{TODO:}
\usage{
ansbacherParser(data, fileType)
}
\arguments{
\item{data}{TODO}

\item{fileType}{TODO}
}
\value{
TODO
}
\description{
This is the description
}
