% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansbacherCommons.R
\name{ansbacherResourcePreemble}
\alias{ansbacherResourcePreemble}
\title{TODO:}
\usage{
ansbacherResourcePreemble(records, accounts, resources, session, ...)
}
\arguments{
\item{records}{TODO}

\item{accounts}{TODO}

\item{resources}{TODO}

\item{session}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
This is the description
}
