% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{appendCash}
\alias{appendCash}
\title{TODO}
\usage{
appendCash(orderedHoldings, nestedHoldings)
}
\arguments{
\item{orderedHoldings}{TODO.}

\item{nestedHoldings}{TODO.}
}
\value{
TODO.
}
\description{
TODO
}
