% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{appendPortfolioIDs}
\alias{appendPortfolioIDs}
\title{This function syncs portfolios from a data frame and appends the name and id's to the data frame}
\usage{
appendPortfolioIDs(data, session, prefix = "")
}
\arguments{
\item{data}{The data frame. Required columns: "portfolio" ("name of portoflio"),}

\item{session}{The rdecaf session}

\item{prefix}{Optional string to be used as prefix for portfolio name:}
}
\value{
Returns the appended data frame.
}
\description{
This is the description
}
