% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionIncrement.R
\name{appendStocks}
\alias{appendStocks}
\title{This function masks decaf positions based on provider positions:}
\usage{
appendStocks(ePos, stocks, pDate, resources)
}
\arguments{
\item{ePos}{The excess position from provider.}

\item{stocks}{The stocks on Decaf.}

\item{pDate}{The position date.}

\item{resources}{The resources data frame}
}
\value{
Return the stocks with the excess provider position appended.
}
\description{
This is the description
}
