% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{areaTimeSeriesPlot}
\alias{areaTimeSeriesPlot}
\title{Provides an area time series plot.}
\usage{
areaTimeSeriesPlot(df, smooth = 0.3, limitFactor = 0.1, title,
  ylab = "", col = "darkslategray3", ylabFun = NULL)
}
\arguments{
\item{df}{A data frame with 'value' and 'date' columns.}

\item{smooth}{The smoothing parameter.}

\item{limitFactor}{The factor with which to expand the limits.}

\item{title}{The title of the plot.}

\item{ylab}{The y-axis name.}

\item{col}{The color for the polygon.}

\item{ylabFun}{The transformation function to apply to the ylab values.}
}
\value{
An area time series plot.
}
\description{
This is the description
}
