% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{argsParser}
\alias{argsParser}
\title{A function to parse command line arguments}
\usage{
argsParser(commandArgs)
}
\arguments{
\item{commandArgs}{Command line arguments whereby arguments sets are separated by --}
}
\value{
A list with the arguments:
}
\description{
A function to parse command line arguments
}
