% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{arrangeHoldings}
\alias{arrangeHoldings}
\title{Arranges holdings based on given columns.}
\usage{
arrangeHoldings(holdings, col)
}
\arguments{
\item{holdings}{The holdings data frame.}

\item{col}{The columns to be arranged by.}
}
\value{
A list.
}
\description{
This is a description.
}
