% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assertit}
\alias{assertit}
\title{Stop with error if condition is FALSE.}
\usage{
assertit(condition, error)
}
\arguments{
\item{condition}{Boolean value.}

\item{error}{Error message.}
}
\value{
Nothing but exception!
}
\description{
Stop with error if condition is FALSE.
}
