% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{assetPerformanceWrapper}
\alias{assetPerformanceWrapper}
\title{A wrapper function for calculating the position wise performance and pnls.}
\usage{
assetPerformanceWrapper(portfolio, date, priceField, periods, session,
  includeStats = FALSE, excludeWeekends = TRUE, holdings = NULL,
  hParams = NULL, resources = NULL)
}
\arguments{
\item{portfolio}{The portfolio id.}

\item{date}{The date of the report.}

\item{priceField}{The name of the price id column.}

\item{periods}{The desired periods.}

\item{session}{The rdecaf session.}

\item{includeStats}{Should advanced stats be included?}

\item{excludeWeekends}{Should weekends be excluded from consideration?}

\item{holdings}{If pnls are desired, holdings mus be supplied. Default is NULL.}

\item{hParams}{If pnls are desired, holdings params must be supplied. Default is NULL.}

\item{resources}{The resources data frame. If NULL (Default), function will get.}
}
\value{
A list with asset returns, asset statistics and asset pnls.
}
\description{
This is the description
}
