% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{barPlot}
\alias{barPlot}
\title{Provides bar plot.}
\usage{
barPlot(df, title, colors = "darkslategray3")
}
\arguments{
\item{df}{A data frame with columns 'value' and 'date.}

\item{title}{The title of the plot.}

\item{colors}{The color(s).}
}
\value{
A bar plot.
}
\description{
This is the description
}
