% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{barPnl}
\alias{barPnl}
\title{This function generates a PnL bar chart given a numerical y value (e.g. drawdown) and categorical x value (e.g. month).}
\usage{
barPnl(df, xval, yval, fillval, title, xtitle, just = -0.1, size = 3)
}
\arguments{
\item{df}{the data frame containing the x and y values.}

\item{xval}{the x value as a string.}

\item{yval}{the y value as a string.}

\item{fillval}{the column to use to fill the color of the bars.}

\item{title}{the title given to the entire plot.}

\item{xtitle}{the title given to the x axis.}

\item{just}{the value for the justification of the labels, has a default.}

\item{size}{the value for the size of the labels, has a default.}
}
\value{
A bar plot.
}
\description{
This is the description
}
