% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{barRet}
\alias{barRet}
\title{This function generates a returns plots given a series.}
\usage{
barRet(series, ttl = "Weekly Returns")
}
\arguments{
\item{series}{a series of returns, e.g. weekly.}

\item{ttl}{the title given to the plot.}
}
\value{
A bar plot with gradient color for positive and negative returns.
}
\description{
This is the description
}
