% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbg.R
\name{bbgTransformFuns}
\alias{bbgTransformFuns}
\title{A function to provide the transformations for bbg results.}
\usage{
bbgTransformFuns()
}
\value{
A list with the memnonic as the name of the list and the function as the element.
}
\description{
A function to provide the transformations for bbg results.
}
