% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbg.R
\name{bdlhs}
\alias{bdlhs}
\title{R wrapper around bdlhs.}
\usage{
bdlhs(host, port, user, pass, flds, tcks, exe = "bdlhs")
}
\arguments{
\item{host}{Remote SFTP server hostname.}

\item{port}{Remote SFTP server port.}

\item{user}{Remote SFTP username.}

\item{pass}{Remote SFTP password.}

\item{flds}{The fields.}

\item{tcks}{The tickers.}

\item{exe}{Path to or name of `bdlhs` executable.}
}
\value{
A data frame
}
\description{
R wrapper around bdlhs.
}
