% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{beanbagNAVTable}
\alias{beanbagNAVTable}
\title{Prepares the NAV data frame.}
\usage{
beanbagNAVTable(x, inception, pxinfo, investments = NULL,
  useIntYtd = FALSE)
}
\arguments{
\item{x}{A list with "portfolio", "ccy", "currentNAV", "previous NAV".}

\item{inception}{The launch date.}

\item{pxinfo}{The YTD Performance.}

\item{investments}{The investment/transfer trades.}

\item{useIntYtd}{Should the internal YTD be used if external doesn't exist? Default is FALSE.}
}
\value{
A data frame.
}
\description{
This is a description.
}
