% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{beanbagPerformanceTable}
\alias{beanbagPerformanceTable}
\title{Prepares the performance table as styled html.}
\usage{
beanbagPerformanceTable(stats)
}
\arguments{
\item{stats}{The statistics data frame from the endpoint.}
}
\value{
A styled html table.
}
\description{
This is a description.
}
