% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{beanbagStringHighlight}
\alias{beanbagStringHighlight}
\title{Highlights each word in a string with a background color}
\usage{
beanbagStringHighlight(string, textColor = "white",
  bgColor = "darkblue", fontSize = 8.5)
}
\arguments{
\item{string}{A vector with strings.}

\item{textColor}{The text color.}

\item{bgColor}{The background color.}

\item{fontSize}{The font size.}
}
\value{
An html object.
}
\description{
This is a description.
}
