% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{beautify}
\alias{beautify}
\title{A Wrapper function for formatting numbers:}
\usage{
beautify(number, nsmall = 0)
}
\arguments{
\item{number}{An array of numeric values.}

\item{nsmall}{The number of digits.}
}
\value{
Formated number(s):
}
\description{
This is a description.
}
