% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{benchmarkAnalysis}
\alias{benchmarkAnalysis}
\title{A function used in computeReturnStats for benchmark metrics, e.g. treynor.}
\usage{
benchmarkAnalysis(data, minobs = 30)
}
\arguments{
\item{data}{a df with ts columns (benchmark and container column names)}

\item{minobs}{min required obs to run the advanced statitistics. defaults to 30}
}
\value{
A list with key metrics.
}
\description{
This is the description
}
