% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueoceanCommons.R
\name{blueoceanPosColumnMappers}
\alias{blueoceanPosColumnMappers}
\title{A function to return the column mappers}
\usage{
blueoceanPosColumnMappers()
}
\value{
Returns a list with the column mappers.
}
\description{
This is the description
}
