% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueoceanCommons.R
\name{blueoceanSecTypeMapper}
\alias{blueoceanSecTypeMapper}
\title{A function to return the security type mapper}
\usage{
blueoceanSecTypeMapper()
}
\value{
Returns a list with the security type mappers.
}
\description{
This is the description
}
