% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{bondMaturityFromTicker}
\alias{bondMaturityFromTicker}
\title{A function to infer the maturity of bond from bloomberg ticker:}
\usage{
bondMaturityFromTicker(ticker)
}
\arguments{
\item{ticker}{The bloomberg ticker.}
}
\value{
Returns the inferred maturity
}
\description{
This is a description.
}
