% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeorder.R
\name{bookExecutions}
\alias{bookExecutions}
\title{A function to book executions from incoming external execution data.}
\usage{
bookExecutions(resources, trdOrdExecutions, trdOrds, allocationField,
  trdOrdAllocations, data, bookingFun, session)
}
\arguments{
\item{resources}{The rdecaf resources data-frame.}

\item{trdOrdExecutions}{The trade order executions data-frame.}

\item{trdOrds}{The trade order data-frame.}

\item{allocationField}{The name of the allocation id column in the incoming execution data.}

\item{trdOrdAllocations}{The trade order allocation data-frame.}

\item{data}{The incoming external execution data.}

\item{bookingFun}{The booking function for the incoming execution data.}

\item{session}{The rdecaf session.}
}
\description{
This is the description
}
