% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capitalise}
\alias{capitalise}
\title{A function to capitalise a string.}
\usage{
capitalise(str)
}
\arguments{
\item{str}{A vector with strings.}
}
\value{
Capitalised strings.
}
\description{
A function to capitalise a string.
}
