% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{centeredColorScale}
\alias{centeredColorScale}
\title{Prepares a data frame with scaled hex color codes for values centerd around zero.}
\usage{
centeredColorScale(
  values,
  positiveRange = c("lightgreen", "darkgreen"),
  negativeRange = c("orangered3", "coral"),
  stepsno = 15
)
}
\arguments{
\item{values}{A vector with values centered around zero.}

\item{positiveRange}{A vector with 2 strings with the range of colors for positive values.}

\item{negativeRange}{A vector with 2 strings with the range of colors for negative values.}

\item{stepsno}{The number of steps for the color range.}
}
\value{
A data frame with the values and the scaled hex color codes.
}
\description{
This is a description.
}
