% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkCondition}
\alias{checkCondition}
\title{A function to check condition of numeric value:}
\usage{
checkCondition(vector, val, condition = "equal")
}
\arguments{
\item{vector}{The numeric value to check condition for.}

\item{val}{The numeric value to check condition against>}

\item{condition}{Either of 'equal', 'greater', 'smaller'.}
}
\value{
Returns either NA, TRUE or FALSE.
}
\description{
This is a description.
}
