% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classify}
\alias{classify}
\title{A function to transform each column in data frame to
the correct class.}
\usage{
classify(df)
}
\arguments{
\item{df}{The data frame.}
}
\value{
The data frame with correct column classes.
}
\description{
A function to transform each column in data frame to
the correct class.
}
