% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanDfReturn}
\alias{cleanDfReturn}
\title{This function removes undesired columns from a returned data frame.}
\usage{
cleanDfReturn(df, col = "id")
}
\arguments{
\item{df}{the data frame.}

\item{col}{the undesired column string. Defaults to id.}
}
\value{
a cleaned data frame.
}
\description{
This is a description.
}
