% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav.R
\name{cleanOutlier}
\alias{cleanOutlier}
\title{Given an outliers data frame, attempts to explain them using DECAF transaction data}
\usage{
cleanOutlier(
  outliers,
  session,
  portfolio = NULL,
  date = NULL,
  currency = NULL,
  excluded = NULL,
  inferred = NULL,
  correct = 0.5,
  thresh = 0.02,
  toler = 0.05
)
}
\arguments{
\item{outliers}{the outliers data frame}

\item{session}{The rdecaf session}

\item{portfolio}{The portfolio id. Not required, defaults to NULL, taken from outliers if not found}

\item{date}{The outlier date. Not required, defaults to NULL, taken from outliers if not found}

\item{currency}{The outlier currency. Not required, defaults to NULL, taken as reference currency if not found}

\item{excluded}{vector defining resources (from quants endpoint) to exclude in analysis, e.g. fx future. Defaults to NULL}

\item{inferred}{vector defining resources (from quants endpoint) to switch the sign of, e.g. loan. Defaults to NULL}

\item{correct}{magnitude factor applied in calculation used to derive relevancy of outlier residual.}

\item{thresh}{magnitude factor applied in calculation used to derive relevancy of outlier residual.}

\item{toler}{buffer factor applied in calculation used to derive relevancy of outlier residual.}
}
\value{
A data frame with the relevant associated transaction returned for the outliers inout.
}
\description{
This is the description
}
