% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clever.R
\name{clever_annotate}
\alias{clever_annotate}
\title{A function to generate the clever message syntax.}
\usage{
clever_annotate(msg)
}
\arguments{
\item{msg}{A message as string.}
}
\description{
This is the description
}
