% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clever.R
\name{clever_example}
\alias{clever_example}
\title{DECAF Clever job example.}
\usage{
clever_example(session, fail = FALSE, ...)
}
\arguments{
\item{session}{`rdecaf` session.}

\item{fail}{If `TRUE`, the function will stop with an error message.}

\item{...}{Further arguments to be ignored.}
}
\description{
This example attempts to log some DECAF Instance information that is
retrieved from DECAF Instance API.
}
