% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clever.R
\name{clever_stop}
\alias{clever_stop}
\title{Issues a DECAF Clever annotation output message and stops the
script with the same message.}
\usage{
clever_stop(msg)
}
\arguments{
\item{msg}{Annotation message.}
}
\description{
Issues a DECAF Clever annotation output message and stops the
script with the same message.
}
