% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnl.R
\name{compilePnlReport}
\alias{compilePnlReport}
\title{Compiler fn that returns all elements to construct RELATIVE pnl report.}
\usage{
compilePnlReport(portfolio, until = Sys.Date(), session)
}
\arguments{
\item{portfolio}{the decaf portfolio id.}

\item{until}{the end date for the pnl relative period calculation.}

\item{session}{the decaf session.}
}
\value{
A list containing the relative period PnL dfs and nav values.
}
\description{
This is the description
}
