% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{complianceBreachInspector}
\alias{complianceBreachInspector}
\title{A function to inspect the breaches in the latest compliance checks.}
\usage{
complianceBreachInspector(resources, session, occurThreshold = 2)
}
\arguments{
\item{resources}{The resources data frame.}

\item{session}{The rdecaf session.}

\item{occurThreshold}{For common instrument culprits, what is the occurance threshold? Default = 2.}
}
\value{
A list with the inspected data.
}
\description{
A function to inspect the breaches in the latest compliance checks.
}
