% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{complianceBreachReport}
\alias{complianceBreachReport}
\title{Returns the breach report as a list with the aggregate, standalone and report elements.}
\usage{
complianceBreachReport(
  portfolios,
  session,
  divisor = "nav",
  field,
  minLim,
  maxLim,
  rDate = Sys.Date(),
  rName = "Breach Report",
  excludeCtypes = NULL
)
}
\arguments{
\item{portfolios}{The portfolio id's.}

\item{session}{The session info.}

\item{divisor}{The consolidation list name to be used as the divisor. Default is nav.}

\item{field}{The resources field. i.e issuer for issuer concentration or id for single concentration.}

\item{minLim}{The lower limit.}

\item{maxLim}{The upper limit.}

\item{rDate}{The date of the report. Default is Sys.Date.}

\item{rName}{The name of the report. Default is Breach Report.}

\item{excludeCtypes}{A vector with strings, i.e CCY or NULL. Default is NULL.}
}
\value{
A list.
}
\description{
This is a description.
}
