% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{COMPOSITION2}
\alias{COMPOSITION2}
\title{A function to define a fund report composition}
\usage{
COMPOSITION2(portfolio, ccy, date, dtype, regions, session)
}
\arguments{
\item{portfolio}{The portfolio id}

\item{ccy}{The currency of reporting}

\item{date}{The date of reporting}

\item{dtype}{The date type of reporting}

\item{regions}{The list with the country to region mapping}

\item{session}{The rdecaf session}
}
\value{
A data-frame with the printable format of holdings
}
\description{
This is a description.
}
