% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{computeAutomationRate}
\alias{computeAutomationRate}
\title{This function computes the automation rate for users.}
\usage{
computeAutomationRate(userActions, includeAUTO = TRUE)
}
\arguments{
\item{userActions}{The data frame with the use actions.}

\item{includeAUTO}{a logical parameter.}
}
\value{
A table with the automation rate.
}
\description{
This function computes the automation rate for users.
}
