% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{computeCurrencyExposure}
\alias{computeCurrencyExposure}
\title{This function computes the ytd monthly returns using the pconsolidation's internal navs.}
\usage{
computeCurrencyExposure(holdings, ccy)
}
\arguments{
\item{holdings}{The holdings data frame from getFlatHoldings.}

\item{ccy}{The portfolio currency.}
}
\value{
The fx forward hedge adjusted currency exposure
}
\description{
This function computes the ytd monthly returns using the pconsolidation's internal navs.
}
