% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnl.R
\name{computeEvents}
\alias{computeEvents}
\title{Computes the PnL from the contextualized data.}
\usage{
computeEvents(context)
}
\arguments{
\item{context}{the list containing contextualized data frame ledgers.}
}
\value{
A list containing the ledger artifact info, the clean ledger data, the summarised one liner data, a dubious flag, and the original ledger data.
}
\description{
This is the description
}
