% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbg.R
\name{computeFXCrosses}
\alias{computeFXCrosses}
\title{A function to compute FX crosses. It assume 'USD' as anchor currency and that inverses of rate for anchor vs minor is provided.}
\usage{
computeFXCrosses(allPairs, pricedPairs)
}
\arguments{
\item{allPairs}{A data frame with all the currency pairs. Requires "main", "altn" and "symbol" columns.}

\item{pricedPairs}{A data frame with the priced currency pairs. Requires "symbol", "close" and "date" columns.}
}
\value{
A list with the ohlc observations and invalid symbols.
}
\description{
A function to compute FX crosses. It assume 'USD' as anchor currency and that inverses of rate for anchor vs minor is provided.
}
