% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{computePeriodicReturns}
\alias{computePeriodicReturns}
\title{A function to compute the periodic returns of xts price.}
\usage{
computePeriodicReturns(price)
}
\arguments{
\item{price}{An xts price vector.}
}
\value{
A xts returns data frame.
}
\description{
A function to compute the periodic returns of xts price.
}
