% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnl.R
\name{computePnL}
\alias{computePnL}
\title{Computes the pnl using the quant context list.}
\usage{
computePnL(quantContext)
}
\arguments{
\item{quantContext}{A list with the quant context.}
}
\value{
A list with the quant context extended by pnl computation.
}
\description{
This is the description
}
