% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{computeReturnStats}
\alias{computeReturnStats}
\title{Provides the return statistics for a data frame with price and date column}
\usage{
computeReturnStats(
  df,
  pxCol,
  dtCol,
  method = "discrete",
  returnOnly = FALSE,
  benchmark = NULL,
  rfts = NA
)
}
\arguments{
\item{df}{The data frame.}

\item{pxCol}{The name of the price column.}

\item{dtCol}{The name of the date column.}

\item{method}{The return calcuation method:'discrete', 'log'}

\item{returnOnly}{Should only the Total Return be calculated?}

\item{benchmark}{The benchmark symbol. Defaul NULL.}

\item{rfts}{The the risk free time series. Defaults NULL.}
}
\value{
A data frame with the return statistics.
}
\description{
This is the description
}
