% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{conditionFn}
\alias{conditionFn}
\title{This function returns a given data frame with a logical column based on a supplied condition}
\usage{
conditionFn(df, expr, group = NULL)
}
\arguments{
\item{df}{the data frame.}

\item{expr}{string specifying the logic used to derive the 'condition' column.}

\item{group}{string vector of variable names in case any derivation should be grouped. Defaults to NULL.}
}
\value{
the supplied df with a new column called 'condition'.
}
\description{
This is a description.
}
