% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnl.R
\name{contextEvents}
\alias{contextEvents}
\title{Contextualizes the flattened PnL data.}
\usage{
contextEvents(flat, excludedTags = c("exclude", "pnl"))
}
\arguments{
\item{flat}{the list containing clean data frame ledgers.}

\item{excludedTags}{the tags to exclude from context in the ledger df.}
}
\value{
A list containing the contextualized ledger data frame and the original one.
}
\description{
This is the description
}
