% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnl.R
\name{contextEvents__}
\alias{contextEvents__}
\title{Contextualizes the flattened PnL data.}
\usage{
contextEvents__(flat, excludedTags = c("exclude", "pnl"), session = NULL)
}
\arguments{
\item{flat}{the list containing clean data frame ledgers.}

\item{excludedTags}{the tags to exclude from context in the ledger df.}

\item{session}{the decaf session. Defaults to NULL TODO.}
}
\value{
A list containing the contextualized ledger data frame and the original one.
}
\description{
This is the description
}
