% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnl.R
\name{contextualizeQuants}
\alias{contextualizeQuants}
\title{A function to contextualize quants using the pnl preemble.}
\usage{
contextualizeQuants(pnlPreemble, dateBeg, dateEnd)
}
\arguments{
\item{pnlPreemble}{A list with the extended positions and quants.}

\item{dateBeg}{The beginning date.}

\item{dateEnd}{The ending date.}
}
\value{
A list with the quants contexts for each position.
}
\description{
This is the description
}
