% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{countryNormalizer}
\alias{countryNormalizer}
\title{A function to normalise country names.}
\usage{
countryNormalizer(str)
}
\arguments{
\item{str}{A string vector.}
}
\value{
The vector with the normalised country names.
}
\description{
This is a description.
}
