% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{countryRegionMap}
\alias{countryRegionMap}
\title{A function the country to region mapping}
\usage{
countryRegionMap()
}
\value{
Returns a list with the country to region mapping.
}
\description{
This is a description.
}
