% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{createAssetClasses}
\alias{createAssetClasses}
\title{Creates the asset class hierarcy}
\usage{
createAssetClasses(aclassConfig, session)
}
\arguments{
\item{aclassConfig}{The list with the asset class configuration.}

\item{session}{The rdecaf session.}
}
\value{
Creates the asset classes and returns NULL
}
\description{
This is the description
}
