% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createBatches}
\alias{createBatches}
\title{A function to provide the start and end indices of desired batches.}
\usage{
createBatches(n, batchSize)
}
\arguments{
\item{n}{The length}

\item{batchSize}{The desired batch size}
}
\value{
A list with starting and ending indices
}
\description{
A function to provide the start and end indices of desired batches.
}
