% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{createFXFutureResource}
\alias{createFXFutureResource}
\title{A function to create fx forward resources:}
\usage{
createFXFutureResource(df, session)
}
\arguments{
\item{df}{The data-frame with the resource information}

\item{session}{The rdecaf session.}
}
\value{
Returns the http post result.
}
\description{
This is a description.
}
