% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohlc.R
\name{createOhlcIndexSeriesFromReturns}
\alias{createOhlcIndexSeriesFromReturns}
\title{This function create a new index price series from returns in the ohlc table.}
\usage{
createOhlcIndexSeriesFromReturns(
  session,
  sourceKeyword = "\\\\|Return",
  targetReplacement = "|PXLAST",
  excludeWeekends = FALSE
)
}
\arguments{
\item{session}{The rdecaf session.}

\item{sourceKeyword}{The keyword to look for in the ohlcs denoting returns}

\item{targetReplacement}{The replacement of the sourceKeyword to be used for the new ohlc series}

\item{excludeWeekends}{Parameter for getOhlcObsForSymbol. Shall weekends be excluded. Default is FALSE.}
}
\description{
This is a description
}
