% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{createReferenceTrade}
\alias{createReferenceTrade}
\title{A function to create a trade reference.}
\usage{
createReferenceTrade(date, account, resource, qty, px, ext = NULL)
}
\arguments{
\item{date}{The trade date.}

\item{account}{The account id.}

\item{resource}{The resource id.}

\item{qty}{The quantity of the trade.}

\item{px}{The price of the trade.}

\item{ext}{A custom extension. Default is NULL.}
}
\value{
A reference string.
}
\description{
This is the description
}
