% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{createResourcesWrapper}
\alias{createResourcesWrapper}
\title{A function to match instrument in a data frame with resources in the system.}
\usage{
createResourcesWrapper(data, session)
}
\arguments{
\item{data}{A data-frame with the resource information.}

\item{session}{The rdecaf session.}
}
\value{
The data-frame with the resmain column
}
\description{
This is a description.
}
