% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{ctype_mapping_figi_to_decaf}
\alias{ctype_mapping_figi_to_decaf}
\title{A function to return the figi to decaf ctype file mappers.}
\usage{
ctype_mapping_figi_to_decaf()
}
\value{
Returns a list with the mapping.
}
\description{
This is the description
}
