% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohlc.R
\name{customOhlcRanked}
\alias{customOhlcRanked}
\title{This function create a new custom ohlc series and observations based on existing ohlc oberservations
by stichting them based on the order/rank.}
\usage{
customOhlcRanked(resources, session, lookBack = 60, constituents)
}
\arguments{
\item{resources}{The resources data frame.}

\item{session}{The rdecaf session.}

\item{lookBack}{The lookBak period.}

\item{constituents}{A list with the existing ohlc symbols and the desired target symbol such as: list("ohlc1"="XYZ|1", "ohlc2"=XYZ|2, target=XYZ|Custom)}
}
\description{
This is a description
}
