% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dateOfPeriod}
\alias{dateOfPeriod}
\title{A function to get the date of a specific verbosely expressed date (expressed by memnonic).}
\usage{
dateOfPeriod(memnonic = "D-0")
}
\arguments{
\item{memnonic}{Follow the convention: [Period Letter] - [lookback], e.g "D-0" or "W-3".}
}
\value{
The specific date of the date memnonic.
}
\description{
This is a description
}
