% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remapDB.R
\name{dbRemapBBGTickerCleaner}
\alias{dbRemapBBGTickerCleaner}
\title{This function cleans possible errors in bloomberg ticker symbols.}
\usage{
dbRemapBBGTickerCleaner(tickers)
}
\arguments{
\item{tickers}{A vector with bloomberg ticker symbols.}
}
\value{
Returns a vector with cleaned bloomberg ticker symbols.
}
\description{
This is a description.
}
