% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remapDB.R
\name{dbRemapCheckUpdate}
\alias{dbRemapCheckUpdate}
\title{This TODO}
\usage{
dbRemapCheckUpdate(resources, mktProvider, tickerField, dbRemapName,
  isinFlds, idFlds, figiApiKey, figiMField, figiPutMap, figiPatchMap)
}
\arguments{
\item{resources}{TODO}

\item{mktProvider}{TODO}

\item{tickerField}{TODO}

\item{dbRemapName}{TODO}

\item{isinFlds}{TODO}

\item{idFlds}{TODO}

\item{figiApiKey}{TODO}

\item{figiMField}{TODO}

\item{figiPutMap}{TODO}

\item{figiPatchMap}{TODO}
}
\value{
TODO
}
\description{
This is a description.
}
