% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remapDB.R
\name{dbRemapCountryTreater}
\alias{dbRemapCountryTreater}
\title{This function converts letter code or arbitrary country names to full names:}
\usage{
dbRemapCountryTreater(countries, countrymaps)
}
\arguments{
\item{countries}{A string vector with country names & codes.}

\item{countrymaps}{The name of the countryMap content.}
}
\value{
Returns a vector with the correct country names where possible.
}
\description{
This is a description.
}
