% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remapDB.R
\name{dbRemapNewResource}
\alias{dbRemapNewResource}
\title{This function checks if the incoming resource identifer is new and valid}
\usage{
dbRemapNewResource(remapDB, sourceDB, isinFlds = NULL, idFlds = NULL)
}
\arguments{
\item{remapDB}{The remap database as data-frame.}

\item{sourceDB}{The source database as data-frame.}

\item{isinFlds}{The isin field mapper for source database.}

\item{idFlds}{The field mapper for identifier in source database.}
}
\value{
Returns list of list. 1: Are identifiers new? 2: Are identifiers valid?
}
\description{
This is a description.
}
