% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remapDB.R
\name{dbRemapPrepareResources}
\alias{dbRemapPrepareResources}
\title{A function to prepare DECAF resources for dbRemap action.}
\usage{
dbRemapPrepareResources(
  resources,
  exCtypes = c("FXFWD", "CCY", "DEPO", "LOAN", "FWD", "OTHER"),
  rplKeysInField = list(symbol = c(` @` = "@"), symbol = c(` - LONG` = ""), symbol =
    c(` - SHORT` = "")),
  tickerField = "symbol",
  tickerProvider = "BBG"
)
}
\arguments{
\item{resources}{The data-frame with the DECAF resources.}

\item{exCtypes}{The ctypes to be excluded from dbRemap consideration.}

\item{rplKeysInField}{A list with fields and grep/replace keys.}

\item{tickerField}{A string indicating with resource field corresponds to ticker.}

\item{tickerProvider}{A string indicating which market data provider ticker to be used. Currently only "BBG". Default is "BBG".}
}
\value{
A data-frame with the prepared DECAF resources.
}
\description{
This is a description.
}
