% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{decafNAVComparison}
\alias{decafNAVComparison}
\title{A function to compare NAV's between 2 decaf systems.}
\usage{
decafNAVComparison(
  accounts,
  type,
  ccy,
  date,
  sSession,
  tSession,
  tDeployment,
  sDeployment,
  charLimit = 100
)
}
\arguments{
\item{accounts}{The account mapping list.}

\item{type}{The type of container. Either 'accounts' or 'portfolios'.}

\item{ccy}{The currency to be valued.}

\item{date}{The date of the consolidations.}

\item{sSession}{The source session.}

\item{tSession}{The target session.}

\item{tDeployment}{The name of the target deployment.}

\item{sDeployment}{The name of the source deployment.}

\item{charLimit}{The character limit for names.}
}
\value{
A data frame with the NAv comparisons.
}
\description{
A function to compare NAV's between 2 decaf systems.
}
