% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{decafPushVouchers}
\alias{decafPushVouchers}
\title{Syncs vouchers in batches.}
\usage{
decafPushVouchers(vouchers, session, N = 500)
}
\arguments{
\item{vouchers}{The voucher data frame.}

\item{session}{The rdecaf session.}

\item{N}{The batch size.}
}
\description{
This is the description
}
