% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{decafSyncOHLC}
\alias{decafSyncOHLC}
\title{This function syncs the ohlcs observations between source DECAF and target DECAF.}
\usage{
decafSyncOHLC(
  sSession,
  tSession,
  resources,
  lte = NULL,
  lookBack = NULL,
  ohlccodeKey = "~DEBSM"
)
}
\arguments{
\item{sSession}{The target session.}

\item{tSession}{The source session.}

\item{resources}{The resources data frame at target.}

\item{lte}{The less than or equal to date for ohlc observations. Default=NULL}

\item{lookBack}{The lookback period. Default=NULL.}

\item{ohlccodeKey}{The key for ohlccode to be used as filter.}
}
\value{
A data frame with ohlc observations.
}
\description{
This is the description
}
