% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{decafSyncResources}
\alias{decafSyncResources}
\title{This function syncs the resources between source DECAF and target DECAF.}
\usage{
decafSyncResources(tSession, sSession, sAccountNames, overrideTypes = NULL)
}
\arguments{
\item{tSession}{The target session.}

\item{sSession}{The source session.}

\item{sAccountNames}{The account names at source.}

\item{overrideTypes}{A vector with strings defining the resource ctypes to be overriden. Default NULL}
}
\value{
The target resources data frame.
}
\description{
This is the description
}
