% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{deleteResourcesByIdentifier}
\alias{deleteResourcesByIdentifier}
\title{Delete resources by identifier, either symbol, isin or id.}
\usage{
deleteResourcesByIdentifier(
  identifier,
  idType,
  resources,
  excludeCols = NULL,
  excludeVals = NULL,
  session
)
}
\arguments{
\item{identifier}{A vector with identifier}

\item{idType}{The type of identifier, either symbol, isin or id.}

\item{resources}{The resources data frame.}

\item{excludeCols}{A vector with resource columns to apply a filter}

\item{excludeVals}{A vector with values to filter out from resource column.}

\item{session}{The rdecaf session.}
}
\description{
This is the description
}
