% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{detectTransfers}
\alias{detectTransfers}
\title{A function to flag all suspected premium payments from trades data, for all cash trades.}
\usage{
detectTransfers(
  portfolio,
  fpKeywords = list(stype = "reversal", remarks = "fee"),
  transferMin = 0.035,
  session
)
}
\arguments{
\item{portfolio}{portfolio ID}

\item{fpKeywords}{list of column and contents to identify as false positives}

\item{transferMin}{min threshold decimal of initial premium to count a transfer as a true premium paid}

\item{session}{The rdecaf session}
}
\value{
A data frame
}
\description{
This is the description
}
