% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dfToStrings}
\alias{dfToStrings}
\title{This function converts data frame into matrix with strings}
\usage{
dfToStrings(df)
}
\arguments{
\item{df}{The data frame.}
}
\value{
A matrix with strings.
}
\description{
This is a description.
}
