% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{distPlots}
\alias{distPlots}
\title{This function generates distribution and QQ plots.
It assumes there is an x value that explains a y value (e.g. benchmark values).}
\usage{
distPlots(df, yval, cval, pcol = "#2C3E50", ncol = "#B15928", ttl, ttl2)
}
\arguments{
\item{df}{the data frame containing the x and y values.}

\item{yval}{the y value as a string.}

\item{cval}{the categorical color value as a string that segments returns.}

\item{pcol}{the color of one of the sets of returns.}

\item{ncol}{the color of the other.}

\item{ttl}{the title given to the distribution plot.}

\item{ttl2}{the title given to the QQ plot.}
}
\value{
A list containing two comparative plots.
}
\description{
This is the description
}
