% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbg.R
\docType{data}
\name{.bbgDecafFieldMap}
\alias{.bbgDecafFieldMap}
\title{A list with the mapping between bbg and decaf fields}
\format{An object of class \code{list} of length 4.}
\usage{
.bbgDecafFieldMap
}
\value{
A list
}
\description{
A list with the mapping between bbg and decaf fields
}
\keyword{datasets}
