% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.formatDate}
\alias{.formatDate}
\title{Formats the given ISO string date to european style.}
\usage{
.formatDate(value)
}
\arguments{
\item{value}{ISO string date.}
}
\value{
European string date.
}
\description{
This is a description.
}
