% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{.getFileName}
\alias{.getFileName}
\title{This is a helper function for download documents function.}
\usage{
.getFileName(response, default = NULL)
}
\arguments{
\item{response}{The response.}

\item{default}{NULL.}
}
\value{
The file name
}
\description{
This is the description
}
