% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.replaceNA}
\alias{.replaceNA}
\title{A function to grep the NA and replace with ""}
\usage{
.replaceNA(x, repl = "")
}
\arguments{
\item{x}{A vector}

\item{repl}{The value to replace the NA's with}
}
\value{
A vector with NA's removed.
}
\description{
This is a description.
}
