% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbg.R
\docType{data}
\name{.transBBGResults}
\alias{.transBBGResults}
\title{A list with the tranform functions for bbg results by field.}
\format{An object of class \code{list} of length 14.}
\usage{
.transBBGResults
}
\value{
A list
}
\description{
A list with the tranform functions for bbg results by field.
}
\keyword{datasets}
