% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{download}
\alias{download}
\title{This is a function to download document responses in decaf.}
\usage{
download(..., params = list(), session = NULL)
}
\arguments{
\item{...}{TODO.}

\item{params}{The rdecaf request parameters.}

\item{session}{The rdecaf session.}
}
\value{
The request response.
}
\description{
This is the description
}
