% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{downloadContent}
\alias{downloadContent}
\title{This is a function for download document content function in decaf.}
\usage{
downloadContent(..., params = list(), session = NULL)
}
\arguments{
\item{...}{TODO.}

\item{params}{The rdecaf request parameters.}

\item{session}{The rdecaf session.}
}
\value{
The content of request response.
}
\description{
This is the description
}
