% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{downloadFile}
\alias{downloadFile}
\title{This is a function to download document files in decaf.}
\usage{
downloadFile(
  ...,
  params = list(),
  filename = NULL,
  directory = NULL,
  session = NULL
)
}
\arguments{
\item{...}{TODO.}

\item{params}{The rdecaf request parameters.}

\item{filename}{The file name.}

\item{directory}{The directory.}

\item{session}{The rdecaf session.}
}
\value{
The file path
}
\description{
This is the description
}
