% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health.R
\name{dubiousOHLCcodes}
\alias{dubiousOHLCcodes}
\title{This function checks whether ohlccodes are dubious.}
\usage{
dubiousOHLCcodes(resources, jaccardCoeff)
}
\arguments{
\item{resources}{A data-frame with the resources from rdecaf.}

\item{jaccardCoeff}{The threshold for the jaccard string distance coefficient.}
}
\value{
The resource data-frame with dubious ohlccodes.
}
\description{
This is a description
}
