% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health.R
\name{dubiousSymbolBBG}
\alias{dubiousSymbolBBG}
\title{This function checks whether symbols are dubious w.r.t Bloomberg symbology.}
\usage{
dubiousSymbolBBG(symbol)
}
\arguments{
\item{symbol}{A vector with symbols}
}
\value{
A vector of length symbols with TRUE or FALSE.
}
\description{
This is a description
}
