% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health.R
\name{dubiousTradeDates}
\alias{dubiousTradeDates}
\title{This function checks whether trade dates are dubious.}
\usage{
dubiousTradeDates(trades, backdatedness)
}
\arguments{
\item{trades}{A data-frame with the trades from rdecaf.}

\item{backdatedness}{The number of days after which a trade should have been created.}
}
\value{
The trades data-frame with dubious trade dates.
}
\description{
This is a description
}
