% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{duplicatedIsinInPortfolios}
\alias{duplicatedIsinInPortfolios}
\title{A function to push a payload to a decaf instance.}
\usage{
duplicatedIsinInPortfolios(accounts = NULL, resources, session)
}
\arguments{
\item{accounts}{The data-frame with the rdecaf accounts. If NULL, all accounts of session are considered. Default is NULL.}

\item{resources}{The data-frame of the resources in the decaf instance.}

\item{session}{The rdecaf session.}
}
\value{
A data-frame of positions which are duplicated in terms of isin.
}
\description{
This is the description
}
