% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ellipsify}
\alias{ellipsify}
\title{Shortens the string and attaches ellipsis.}
\usage{
ellipsify(str, charLimit = 30, capitalise = TRUE)
}
\arguments{
\item{str}{A string.}

\item{charLimit}{The maximum number of characters.}

\item{capitalise}{Boolean to indicate if string should be capitalised. Default is TRUE.}
}
\value{
Ellipsified string.
}
\description{
This is a description.
}
