% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{emailAlert}
\alias{emailAlert}
\title{A function to email alerts.}
\usage{
emailAlert(
  df,
  team = NULL,
  emailList,
  emailParams,
  provider = "DECAF",
  caption = NULL,
  sourceType = "API",
  emailBody,
  deployment,
  subject,
  session
)
}
\arguments{
\item{df}{The data frame to be shown as html at bottom of email.}

\item{team}{The name of team. To be provided if team-wise email is chosen. Default=NULL.}

\item{emailList}{The list of email addresses to be sent to.}

\item{emailParams}{The email parameters.}

\item{provider}{The name of the data provider. Default = 'DECAF'.}

\item{caption}{The caption for the table. Default = NULL.}

\item{sourceType}{The name of the source. Descriptive only. Default = API.}

\item{emailBody}{The email body text.}

\item{deployment}{The name of the deployment.}

\item{subject}{The subject of the email.}

\item{session}{The rdecaf session.}
}
\value{
NULL. An email with the alert will be sent.
}
\description{
This is the description
}
