% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbg.R
\name{emailBBGDLPxReport}
\alias{emailBBGDLPxReport}
\title{A function to email the BBGDL price update report.}
\usage{
emailBBGDLPxReport(ohlc, emailContent, emailParams, reportText = NULL,
  session)
}
\arguments{
\item{ohlc}{The list object coming from treatBBGResultsPrice.}

\item{emailContent}{A list with the email contents. Expects "greeting", "deployment", "url".}

\item{emailParams}{A list with the email parameters.}

\item{reportText}{An optional report text to override the default text. Default is NULL.}

\item{session}{The rdecaf session.}
}
\description{
A function to email the BBGDL price update report.
}
