% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeorder.R
\name{emailExecutionReport}
\alias{emailExecutionReport}
\title{A function to email execution & booking report.}
\usage{
emailExecutionReport(trdOrdExecutions, data, fldMap, emailContent, session)
}
\arguments{
\item{trdOrdExecutions}{The trade order executions data-frame.}

\item{data}{The incoming external execution data.}

\item{fldMap}{A list with the field mappings between trade order execution and incoming execution data.}

\item{emailContent}{The content parameters of the email.}

\item{session}{The rdecaf session.}
}
\description{
This is the description
}
