% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{emailHTMLTable}
\alias{emailHTMLTable}
\title{A function to generate email-friendly HTML table.}
\usage{
emailHTMLTable(df, provider, caption, sourceType = "API")
}
\arguments{
\item{df}{A data-frame.}

\item{provider}{The name of the provider of the data to be used in the footer.}

\item{caption}{The caption of the table to be used.}

\item{sourceType}{The type of the source data to be used for the footer.E.g., 'Api'.}
}
\value{
A java-object of class org.apache.commons.mail.SimpleEmail
}
\description{
TODO:
}
