% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{emailReport}
\alias{emailReport}
\title{A function to send email with a custom email body.}
\usage{
emailReport(from, emailList, subject, body, isLocal = FALSE,
  html = TRUE, attachments = NULL)
}
\arguments{
\item{from}{The sender information.}

\item{emailList}{The vector with recipient emal addresses.}

\item{subject}{The subject of the email.}

\item{body}{The body of the email.}

\item{isLocal}{Boolean to indicate whether it is a test environment. Default is FALSE.}

\item{html}{Boolean to indicate whether body is html. Defaul is TRUE.}

\item{attachments}{The paths to attachments.}
}
\value{
A java-object of class org.apache.commons.mail.SimpleEmail
}
\description{
This is the description
}
