% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{excludeRowsWithKeys}
\alias{excludeRowsWithKeys}
\title{A function to exclude rows in which a specified column
has a key.}
\usage{
excludeRowsWithKeys(data, field, keys)
}
\arguments{
\item{data}{The data-frame.}

\item{field}{The field names to be excluded.}

\item{keys}{The keys to look for.}
}
\value{
The data-frame with the excluded rows.
}
\description{
This is a description.
}
