% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeorder.R
\name{executeAllocations}
\alias{executeAllocations}
\title{A function to book executions from incoming external execution data.}
\usage{
executeAllocations(trdOrdExecutions, data, fldMap, session)
}
\arguments{
\item{trdOrdExecutions}{The trade order executions data-frame.}

\item{data}{The incoming external execution data.}

\item{fldMap}{A list with the field mappings between trade order execution and incoming execution data.}

\item{session}{The rdecaf session.}
}
\description{
This is the description
}
