% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{executeValuationAndPool}
\alias{executeValuationAndPool}
\title{This function executes a job give a portfolio id and pools the result until finished.}
\usage{
executeValuationAndPool(
  id = NULL,
  type = "portfolios",
  start = NULL,
  session,
  maxWait = 900
)
}
\arguments{
\item{id}{The container id. Default is NULL, i.e all containers.}

\item{type}{What type container, accounts or portfolios?}

\item{start}{The start date. Default is NULL, i.e from inception.}

\item{session}{The rdecaf session.}

\item{maxWait}{The maximum waiting time in seconds.}
}
\description{
This is the description
}
