% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{exfoliateSeries}
\alias{exfoliateSeries}
\title{Provides the asset returns MTD and YTD including FX returns}
\usage{
exfoliateSeries(series, anchors)
}
\arguments{
\item{series}{The portfolio id.}

\item{anchors}{The anchors}
}
\value{
A data frame with the asset returns.
}
\description{
This is the description
}
