% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expandLimits}
\alias{expandLimits}
\title{This function}
\usage{
expandLimits(x, factor = 0.1)
}
\arguments{
\item{x}{A numeric vector.}

\item{factor}{The factor to expand the limits with.}
}
\value{
A list with the expanded upper and lower limits
}
\description{
This is a description.
}
