% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnl.R
\name{extendPositionByClosedQuants}
\alias{extendPositionByClosedQuants}
\title{Extends the positions data frame from getFlatholdings with opened/closed quants in between position dates.}
\usage{
extendPositionByClosedQuants(quants, positions, resources)
}
\arguments{
\item{quants}{The quants data frame from rdecaf}

\item{positions}{The positions data frame from getFlatHoldings}

\item{resources}{The resources data frame from rdecaf}
}
\value{
The extended positions data frame.
}
\description{
This is the description
}
