% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractToList}
\alias{extractToList}
\title{Extract unique column names to list}
\usage{
extractToList(df, colname)
}
\arguments{
\item{df}{The data frame}

\item{colname}{The column to be extracted}
}
\value{
A list
}
\description{
This is a description.
}
