% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figi.R
\name{figi}
\alias{figi}
\title{A function to get openFigi information on data for given identifier and currency.}
\usage{
figi(data, idType = "ID_ISIN", fld = "isin", ccy = "ccymain",
  figiApi)
}
\arguments{
\item{data}{The data-frame}

\item{idType}{The type of ID to use for openFigi. Default is ID_ISIN.}

\item{fld}{The column name of the data-frame which has the corresponding instrument id.}

\item{ccy}{The column name for the currency of the instrument.}

\item{figiApi}{The api key of the openFigi account.}
}
\value{
Returns a data-frame with openFigi information.
}
\description{
This is a description.
}
