% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figi.R
\name{figiBondTicker}
\alias{figiBondTicker}
\title{This function treats the ticker from figi results:}
\usage{
figiBondTicker(str)
}
\arguments{
\item{str}{A vector with ticker strings.}
}
\value{
A vector with trated figi bond tickers.
}
\description{
This is a description.
}
