% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figi.R
\name{figiCall}
\alias{figiCall}
\title{A function to call the openFigi api for a given openFigi job data-frame.}
\usage{
figiCall(job, apiKey)
}
\arguments{
\item{job}{A data-frame from prepareOpenFigiData.}

\item{apiKey}{The api key of the openFigi account.}
}
\value{
Returns a flattened data-frame with results from the openFigi api call.
}
\description{
This is a description.
}
