% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figi.R
\name{figiResultTreater}
\alias{figiResultTreater}
\title{A function to treat and enrich openFigi information.}
\usage{
figiResultTreater(figiResult)
}
\arguments{
\item{figiResult}{A data-frame from function figiCall.}
}
\value{
Returns a enriched/treated version of the data-frame from figiCall function.
}
\description{
This is a description.
}
