% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figi.R
\name{figiWrapper}
\alias{figiWrapper}
\title{A wrapper function call figi for instruments without resmains in data frame:}
\usage{
figiWrapper(data, idType = "ID_ISIN", fld = "isin", ccy = "ccymain", figiApi)
}
\arguments{
\item{data}{The data-frame}

\item{idType}{The type of ID to use for openFigi. Default is ID_ISIN.}

\item{fld}{The column name of the data-frame which has the corresponding instrument id.}

\item{ccy}{The column name for the currency of the instrument.}

\item{figiApi}{The api key of the openFigi account.}
}
\value{
Returns the original data-frame with figi information appended.
}
\description{
This is a description.
}
