% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav.R
\name{findOutlier}
\alias{findOutlier}
\title{Given a NAV or return time series df, derives the outliers statistically}
\usage{
findOutlier(
  series,
  session,
  fctr = 8,
  flor = 3.5,
  xVal = FALSE,
  portfolio = NULL
)
}
\arguments{
\item{series}{the time series data frame}

\item{session}{The rdecaf session}

\item{fctr}{numeric ceiling factor used in standard deviation calculation to define outlier}

\item{flor}{numeric floor factor used in standard deviation calculation to limit outlier derivation}

\item{xVal}{true/false to dictate whether external valuation overrides should ignore outlier derivation}

\item{portfolio}{The portfolio id. Not required, defaults to NULL}
}
\value{
A data frame with the extreme outliers flagged.
}
\description{
This is the description
}
