% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{findUnderlying}
\alias{findUnderlying}
\title{A function to find underlying instrument symbols for complex instrument, e.g. minifutures, given resources.}
\usage{
findUnderlying(
  resources,
  Ctype = "SPMF",
  keysToRemove = c(" - LONG", " - SHORT"),
  extDataNomenc = "extdata.schedule",
  names = c("Symbol (Minifuture)", "Symbol (Underlying)", "Symbol (Schedule)")
)
}
\arguments{
\item{resources}{A data-frame with the resource information.}

\item{Ctype}{string of the ctype to filter for, defaults to minifutures.}

\item{keysToRemove}{string elements as a vector to remove from symbol field, defaults to long and short.}

\item{extDataNomenc}{the column names to search in for the extra info. Defaults to 'extdata...'}

\item{names}{the column names to apply to the returned DF. Defaults to three symbols representing relevant info for MF.}
}
\value{
The data-frame with the relevant underlying info extracted.
}
\description{
This is a description.
}
