% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figi.R
\name{flatten}
\alias{flatten}
\title{A helper function to flatten the figiCall results:}
\usage{
flatten(x, job)
}
\arguments{
\item{x}{??}

\item{job}{??}
}
\value{
Returns a flattened data-frame from the openFigi POST result.
}
\description{
This is a description.
}
