% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnl.R
\name{flattenEvents}
\alias{flattenEvents}
\title{flattens the pnl endpoint list data from decaf.}
\usage{
flattenEvents(grp)
}
\arguments{
\item{grp}{the returned list of ledgers from function above.}
}
\value{
A list with the flattened pnl ledger data.
}
\description{
This is the description
}
