% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocketBase.R
\name{flattenUni}
\alias{flattenUni}
\title{Pocketbase dependent function to flatten all the endpoints into a master data frame.}
\usage{
flattenUni(rb, paths, eDate = Sys.Date())
}
\arguments{
\item{rb}{The rocketbase authorizaion credentials. Defined externally.}

\item{paths}{The paths of the relevant endpoints, e.g. /api/collections/cfraEquities/records}

\item{eDate}{date filter floor to apply. Defaults to present.}
}
\value{
a master flattened data frame.
}
\description{
This is the description
}
