% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{futureContractSizes}
\alias{futureContractSizes}
\title{A function to provide the future contract size:}
\usage{
futureContractSizes()
}
\value{
Returns a matrix with the future contract size mapping.
}
\description{
This is a description.
}
