% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{futureMonthMap}
\alias{futureMonthMap}
\title{A function to provide the future month memnonic mapper:}
\usage{
futureMonthMap()
}
\value{
Returns a matrix with the future month mapping.
}
\description{
This is a description.
}
