% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{getAccountPortfolioMap}
\alias{getAccountPortfolioMap}
\title{A function to return a flat account to portfolio mapping using the portfolio data-frame}
\usage{
getAccountPortfolioMap(portfolios)
}
\arguments{
\item{portfolios}{The rdecaf portfolio data-frame}
}
\value{
A data-frame with the flat account to portfolio id's
}
\description{
This is the description
}
