% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{getAccountWiseTrades}
\alias{getAccountWiseTrades}
\title{A function to get account-wise trades from a DECAF instance.}
\usage{
getAccountWiseTrades(accounts, session)
}
\arguments{
\item{accounts}{A vector with account id's.}

\item{session}{The DECAF session info.}
}
\value{
A data-frame with DECAF trades.
}
\description{
This is the description
}
