% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{getAggregateExposure}
\alias{getAggregateExposure}
\title{A function to compute the aggregate exposure of holdings data frame.}
\usage{
getAggregateExposure(holdings, keys)
}
\arguments{
\item{holdings}{A data frame with column names.}

\item{keys}{The keys to be aggregated.}
}
\value{
A list with the exposures
}
\description{
A function to compute the aggregate exposure of holdings data frame.
}
