% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdings.R
\name{getAssetClass}
\alias{getAssetClass}
\title{TODO:}
\usage{
getAssetClass(resources)
}
\arguments{
\item{resources}{rdecaf resources.}
}
\value{
A vector with asset class strings.
}
\description{
This is the description
}
