% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{getAssetClasses}
\alias{getAssetClasses}
\title{Get the asset class data frame with the full names appended}
\usage{
getAssetClasses(session)
}
\arguments{
\item{session}{The rdecaf session.}
}
\value{
A data frame with the asset classes.
}
\description{
This is the description
}
