% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{getAssetReturns}
\alias{getAssetReturns}
\title{Provides the asset returns MTD and YTD including FX returns}
\usage{
getAssetReturns(date, ccy, resources, periods, returnOnly, session)
}
\arguments{
\item{date}{The date of consideration.}

\item{ccy}{The portfolio currency.}

\item{resources}{The resource data frame.}

\item{periods}{A vector with periods: c('DTD', 'WTD', 'MTD', 'QTD', 'YTD').}

\item{returnOnly}{Consider only Total Return?}

\item{session}{The rdecaf session.}
}
\value{
A data frame with the asset returns.
}
\description{
This is the description
}
