% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{getBenchmark}
\alias{getBenchmark}
\title{Gets the benchmark elements for a portfolio.}
\usage{
getBenchmark(
  portfolio,
  start,
  end,
  session,
  rfSymbol = "iShares 1-3 Year Treasury Bond ETF"
)
}
\arguments{
\item{portfolio}{The portfolio id.}

\item{start}{The desired start date.}

\item{end}{The desired end date.}

\item{session}{The rdecaf session.}

\item{rfSymbol}{the risk free series, defaults to SHY.}
}
\value{
A list containing all applicable benchmark info.
}
\description{
This is a description.
}
