% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{getConcentrationByResourceField}
\alias{getConcentrationByResourceField}
\title{Returns a data-frame with the concentration of a resource field in a portfolio.}
\usage{
getConcentrationByResourceField(
  portfolio,
  date,
  resources = NULL,
  session,
  field,
  orderBy = "valuation(\%)",
  divField = "nav"
)
}
\arguments{
\item{portfolio}{The portfolio id.}

\item{date}{The date of the report.}

\item{resources}{The resources data frame. If NULL, function will query it. Default NULL.}

\item{session}{The session info.}

\item{field}{The resources field. i.e issuer for issuer concentration or id for single concentration.}

\item{orderBy}{The ordering of the report. Default is valuation(pc).}

\item{divField}{The consolidation list name to be used as the divisor. Default is nav.}
}
\value{
A data frame
}
\description{
This is a description.
}
