% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdings.R
\name{getConsolidationHoldings}
\alias{getConsolidationHoldings}
\title{TODO:}
\usage{
getConsolidationHoldings(session, containerType, containerId, ccy, date,
  charLimit, resources)
}
\arguments{
\item{session}{The rdecaf session}

\item{containerType}{The container type}

\item{containerId}{The container id}

\item{ccy}{The currency of the consolidation}

\item{date}{The date of the consolidation}

\item{charLimit}{The limit of characters of holdings name}

\item{resources}{The rdecaf resources data frame.}
}
\value{
A data-frame with the holdings
}
\description{
This is the description
}
