% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{getCorporateActionPerUnit}
\alias{getCorporateActionPerUnit}
\title{A function to derive the dividend or coupon per unit from decaf's trades endpoint.}
\usage{
getCorporateActionPerUnit(
  session,
  ctype = c("81", "101")[1],
  commitment__gte = dateOfPeriod("Y-0")
)
}
\arguments{
\item{session}{The session of the instance.}

\item{ctype}{The trade ctype. Either 81 or 101. Default is 81.}

\item{commitment__gte}{Since which date should the corporate actions be consdered? When NULL for all history. Default is dateOfPeriod(Y-0).}
}
\value{
A data frame with the corporate action per unit value, resource symbol and id and the corresponding date.
}
\description{
A function to derive the dividend or coupon per unit from decaf's trades endpoint.
}
