% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figi.R
\name{getCurrencyFromFigiExchCode}
\alias{getCurrencyFromFigiExchCode}
\title{A function to infer the currency from an instrument identifier and market sector.}
\usage{
getCurrencyFromFigiExchCode(df)
}
\arguments{
\item{df}{The data-frame with marketSector and idValue columns.}
}
\value{
Returns the original data-frame enriched with the currency.
}
\description{
This is a description.
}
