% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDepName}
\alias{getDepName}
\title{This function returns the deployment string name given session info}
\usage{
getDepName(session)
}
\arguments{
\item{session}{the rdecaf session.}
}
\value{
A character vector of the deployment name.
}
\description{
This is a description.
}
