% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDissimilarRows}
\alias{getDissimilarRows}
\title{Thsi funciton returns rows of a treated data frame which are different compared to the original}
\usage{
getDissimilarRows(altered, original, idField)
}
\arguments{
\item{altered}{The altered data frame.}

\item{original}{The original data frame.}

\item{idField}{The field to be used to align the data frames.}
}
\value{
A data frame with altered rows only.
}
\description{
This is the description
}
