% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDTDSlice}
\alias{getDTDSlice}
\title{A function which provides the DTD slice of the data frame.}
\usage{
getDTDSlice(df, date)
}
\arguments{
\item{df}{The data frame with a 'date' column.}

\item{date}{the report date.}
}
\value{
A data frame with the DTD slice of the data frame or NULL.
}
\description{
This is the description
}
