% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getEndpointLink}
\alias{getEndpointLink}
\title{This function the decaf link embeded in an HTML string given inputs.}
\usage{
getEndpointLink(session, endpnt, df = NULL, placeholder = "LINK")
}
\arguments{
\item{session}{the rdecaf session.}

\item{endpnt}{the rdecaf endpoint, e.g. 'trades'.}

\item{df}{the data frame supplied if there are to be granular links in a detailed URL. Defaults to NULL for the regular site endpoint page.}

\item{placeholder}{the HTMLized text to mask the full URL link. Defaults to 'LINK', but can be the trade ID e.g.}
}
\value{
a character vector of the HTMLized link(s).
}
\description{
This is a description.
}
