% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{getEnrichedStocks}
\alias{getEnrichedStocks}
\title{A function to enrich the stocks data-frame}
\usage{
getEnrichedStocks(stocks, accounts, resources)
}
\arguments{
\item{stocks}{The stocks data-frame.}

\item{accounts}{The accounts data-frame.}

\item{resources}{The resources data-frame.}
}
\value{
Returns a enriched stocks data-frame.
}
\description{
This is a description.
}
