% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbg.R
\name{getExchangeRatePairs}
\alias{getExchangeRatePairs}
\title{A function to build the requestable currency pairs and the entire combinations of currencies}
\usage{
getExchangeRatePairs(mainFX, minorFX)
}
\arguments{
\item{mainFX}{A vector with the main currencies}

\item{minorFX}{A vector with the minor currencies.}
}
\value{
A list with the requestable currency pairs and all combinations.
}
\description{
A function to build the requestable currency pairs and the entire combinations of currencies
}
