% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{getExternalValuation}
\alias{getExternalValuation}
\title{This function gets the external valuations for a portfolio.}
\usage{
getExternalValuation(portfolio, session)
}
\arguments{
\item{portfolio}{The portfolio id.}

\item{session}{The rdecaf session.}
}
\value{
A data frame with the portfolio's external valuations.
}
\description{
This is the description
}
