% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdings.R
\name{getFlatHoldings}
\alias{getFlatHoldings}
\title{A function to flatten the consolidation object.}
\usage{
getFlatHoldings(x, charLimit = 30)
}
\arguments{
\item{x}{The holdings element from getResource("consolidation").}

\item{charLimit}{The character cutoff for the name of the holdings. Default is 50.}
}
\value{
Returns the flat holdings data-frame.
}
\description{
This is the description
}
