% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdings.R
\name{getFormattedHoldings}
\alias{getFormattedHoldings}
\title{A function to generate the formatted holdings data-frame}
\usage{
getFormattedHoldings(holdings)
}
\arguments{
\item{holdings}{The nested holdings.}
}
\value{
Returns the formated holdings data-frame.
}
\description{
This is the description
}
