% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{getFutureExpiry}
\alias{getFutureExpiry}
\title{A function to infer expiry of future bloomberg tickers:}
\usage{
getFutureExpiry(bbgticker, weekday = "Friday", nthWeekday = 3)
}
\arguments{
\item{bbgticker}{The bloomberg ticker.}

\item{weekday}{The day of the week convention.}

\item{nthWeekday}{The nth weekday's date to be used. If weekday is 'Friday' and nthWeekday is '3', it returns the 3rd Friday of the month.}
}
\value{
Returns a date object.
}
\description{
When the bloomberg ticker of a future is provided,
we can infer the date of the expiry.
}
