% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdings.R
\name{getHistoricalPortfolioHoldings}
\alias{getHistoricalPortfolioHoldings}
\title{A function to provide the historical holdings of a portfolio.}
\usage{
getHistoricalPortfolioHoldings(
  portfolio,
  resources = NULL,
  session,
  charLimit = 50,
  regions = NULL,
  start,
  until
)
}
\arguments{
\item{portfolio}{The portfolio id}

\item{resources}{The data-frame as returned by getResource("resources").}

\item{session}{The rdecaf session.}

\item{charLimit}{The character cutoff for the name of the holdings. Default is 50.}

\item{regions}{The data-frame with the country to region mapping.}

\item{start}{The start date.}

\item{until}{The until date.}
}
\value{
The enriched holdings data-frame.
}
\description{
This is the description
}
