% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{getHistoricalUserActions}
\alias{getHistoricalUserActions}
\title{This function gets the historical user actions in a deployment.}
\usage{
getHistoricalUserActions(session, endpoint, exclCols = NULL,
  exclKeys = NULL, actionQueryParams = NULL)
}
\arguments{
\item{session}{The rdecaf session.}

\item{endpoint}{The endpoint to investigate, i.e "trades", "quants".}

\item{exclCols}{The columns to exclude.}

\item{exclKeys}{The keys to exclude from exclCols.}

\item{actionQueryParams}{Addtional query parameters to be passed to params.}
}
\value{
A table with the automation rate.
}
\description{
This function gets the historical user actions in a deployment.
}
