% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{getHoldingsSummary}
\alias{getHoldingsSummary}
\title{Prepares the holdings summary data frame.}
\usage{
getHoldingsSummary(holdings, col = "Subtype", key = "Cash")
}
\arguments{
\item{holdings}{The holdings data frame}

\item{col}{The cash column name}

\item{key}{The cash holdings key}
}
\value{
A list with holdings summary information.
}
\description{
This is a description.
}
