% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdings.R
\name{getHoldingsWrapper}
\alias{getHoldingsWrapper}
\title{A wrapper function to get the enriched holdings based on consolidation parameters.}
\usage{
getHoldingsWrapper(params, resources, session, charLimit = 50, regions = NULL)
}
\arguments{
\item{params}{The parameters for the consolidation.}

\item{resources}{The data-frame as returned by getResource("resources").}

\item{session}{The rdecaf session.}

\item{charLimit}{The character cutoff for the name of the holdings. Default is 50.}

\item{regions}{The data-frame with the country to region mapping.}
}
\value{
The enriched holdings data-frame.
}
\description{
This is the description
}
