% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getIndex}
\alias{getIndex}
\title{A function to check multiple conditions row-wise in data-frame.}
\usage{
getIndex(df, inCols, candCols = NULL, inKeys = NULL, exKeys = NULL,
  inclToIns = NULL)
}
\arguments{
\item{df}{A data-frame.}

\item{inCols}{TODO}

\item{candCols}{TODO}

\item{inKeys}{TODO}

\item{exKeys}{TODO}

\item{inclToIns}{TOOD}
}
\value{
Returns a vector with logicals.
}
\description{
This is a description.
}
