% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{getInternalMonthlyReturns}
\alias{getInternalMonthlyReturns}
\title{This function computes the ytd monthly returns using the pconsolidation's internal navs.}
\usage{
getInternalMonthlyReturns(intPrice, date, session)
}
\arguments{
\item{intPrice}{The internal price series.}

\item{date}{The date of the report.}

\item{session}{The rdecaf session.}
}
\value{
A list with the return objects.
}
\description{
This function computes the ytd monthly returns using the pconsolidation's internal navs.
}
