% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav.R
\name{getInternalNAVShare}
\alias{getInternalNAVShare}
\title{Provides the internal NAV/Share data frame}
\usage{
getInternalNAVShare(portfolio, ccy, date, session)
}
\arguments{
\item{portfolio}{The portfolio id.}

\item{ccy}{The portfolio currency.}

\item{date}{The report date.}

\item{session}{The rdecaf session}
}
\value{
A data frame with the internal nav/share price.
}
\description{
This is the description
}
