% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{getInvestments}
\alias{getInvestments}
\title{This function gets the investments for a portfolio.}
\usage{
getInvestments(portfolio, session)
}
\arguments{
\item{portfolio}{The portfolio id.}

\item{session}{The rdecaf session.}
}
\value{
A data frame with the portfolio's investment trades.
}
\description{
This is the description
}
