% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getLastDayOfPeriod}
\alias{getLastDayOfPeriod}
\title{A function to indicate the last week day within a specific periodicity.}
\usage{
getLastDayOfPeriod(dates, day, period)
}
\arguments{
\item{dates}{A vector with dates.}

\item{day}{The weekday to be considered ("Monday", "Tuesday", ... ,"Friday", etc.)}

\item{period}{The periodicity give in single letter ("W", "M", "Q", "Y").}
}
\value{
The date frame with the lastDayOfPeriod boolean column.
}
\description{
This is the description
}
