% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohlc.R
\name{getLatestPx}
\alias{getLatestPx}
\title{This function returns the last px for an ohlc code as of a given date.}
\usage{
getLatestPx(ohlc, date = Sys.Date(), session)
}
\arguments{
\item{ohlc}{string of the ohlc code to query for in the ohlcobservations endpoint.}

\item{date}{date of the as-of/up-to date to query. Defaults to current day.}

\item{session}{list of the rdecaf session.}
}
\value{
numeric of the most recent PX.
}
\description{
This is a description
}
