% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{getLinkToTradesByDate}
\alias{getLinkToTradesByDate}
\title{TODO:
'
This is the description}
\usage{
getLinkToTradesByDate(portfolio, session, gte, lte = NULL, nojournal = "False")
}
\arguments{
\item{portfolio}{A vector of portfolio id's}

\item{session}{The rdecaf session}

\item{gte}{A date after which the trades should be considered}

\item{lte}{A date before which the trades should be considered. Default is set to gte.}

\item{nojournal}{Either "True" or "False" indicating whether journal entries shall be omitted.}
}
\value{
A vector with string representing the decaf url
}
\description{
TODO:
'
This is the description
}
