% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMonthYearOfCurrentYear}
\alias{getMonthYearOfCurrentYear}
\title{This function produces a vector with the month year for a date's year.
Such as: Jan 19, Feb 19, ... , Dec 19 if date is 2019-06-01.}
\usage{
getMonthYearOfCurrentYear(date)
}
\arguments{
\item{date}{The date.}
}
\value{
A vector with the month year for the date's year in %h %y
}
\description{
This is the description
}
