% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMTDSlice}
\alias{getMTDSlice}
\title{A function which provides the MTD slice of the data frame.}
\usage{
getMTDSlice(df, date)
}
\arguments{
\item{df}{The data frame with a 'date' column.}

\item{date}{the report date.}
}
\value{
A data frame with the MTD slice of the data frame or NULL.
}
\description{
This is the description
}
