% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdings.R
\name{getNestedHoldings}
\alias{getNestedHoldings}
\title{A function to generate a nested holdings data-frame}
\usage{
getNestedHoldings(holdings, levels, toplevel = "Subtype",
  sublevels = c("CCY", "Country"))
}
\arguments{
\item{holdings}{The ordered holdings.}

\item{levels}{TODO}

\item{toplevel}{TODO}

\item{sublevels}{TODO}
}
\value{
Returns the nested holdings data-frame.
}
\description{
This is the description
}
