% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{getNPRTxns}
\alias{getNPRTxns}
\title{A function to get the non-performance related actions for a portfolio.}
\usage{
getNPRTxns(portfolio, pxinfo, date, ccy, session)
}
\arguments{
\item{portfolio}{A vector of portfolio id's}

\item{pxinfo}{The data frame with the pxinfo}

\item{date}{The asof date}

\item{ccy}{The desired ccy.}

\item{session}{The rdecaf session}
}
\value{
A data frame
}
\description{
This is the description
}
