% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{getObjectIDByName}
\alias{getObjectIDByName}
\title{This function returns the id of a data frame with name colum and id columns.}
\usage{
getObjectIDByName(df, name, naVal = 1)
}
\arguments{
\item{df}{The data frame.}

\item{name}{The name to be matched.}

\item{naVal}{The value to be returned if no match.}
}
\value{
A vector with ids.
}
\description{
This is the description
}
