% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohlc.R
\name{getOhlcObsWrapper}
\alias{getOhlcObsWrapper}
\title{This function retrievs the ohlc observations for a give symbol.}
\usage{
getOhlcObsWrapper(symbols, ltes, lookBack, excludeWeekends, session)
}
\arguments{
\item{symbols}{A vector with symbols.}

\item{ltes}{A vector with dates. Less than or equal to this date.}

\item{lookBack}{The number of days to look back from 'lte'. Default is 30 days.}

\item{excludeWeekends}{Should the weekends be excluded? Default is TRUE.}

\item{session}{The rdecaf session.}
}
\value{
A list with result.
}
\description{
This is a description
}
