% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohlc.R
\name{getOHLCSeriesForStocks}
\alias{getOHLCSeriesForStocks}
\title{This function retrieves all OHLC series for stocks in give portfolios.}
\usage{
getOHLCSeriesForStocks(container, containerType, asof, resources, session,
  lookBack = 30, exclType = c("CCY", "DEPO", "LOAN"))
}
\arguments{
\item{container}{The container data-frame from rdecaf (either portfolio or account data-frame).}

\item{containerType}{A string with container type (either "portfolio" or "account").}

\item{asof}{The date of the stocks.}

\item{resources}{The resource data-frame from rdecaf.}

\item{session}{The rdecaf session.}

\item{lookBack}{The lookback period for the ohlc observations.}

\item{exclType}{The instrument types to be excluded from stocks. Default is c("CCY", "DEPO", "LOAN")}
}
\value{
A list with the ohlc observations.
}
\description{
This is a description
}
