% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{getOpenPositionByCtype}
\alias{getOpenPositionByCtype}
\title{A function to prepare a data frame with all positions by ctype and holding period}
\usage{
getOpenPositionByCtype(ctype, resources, portfolios, accounts, session)
}
\arguments{
\item{ctype}{The ctype.}

\item{resources}{The rdecaf resources data frame.}

\item{portfolios}{The rdecaf portfolios data frame.}

\item{accounts}{The rdecaf accounts data frame.}

\item{session}{The rdecaf session.}
}
\value{
A data frame with all holding periods.
}
\description{
This is the description
}
