% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{getOptionsData}
\alias{getOptionsData}
\title{A function to clean up the options df returned from the function above.}
\usage{
getOptionsData(session, date = Sys.Date())
}
\arguments{
\item{session}{list of the rdecaf session.}

\item{date}{date of the as-of/up-to date to query active and unexpired instruments. Defaults to current day.}
}
\value{
The data-frame with identified underlying instruments of options and relevant info, cleaned up for export.
}
\description{
This is a description.
}
