% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{getOptionsFromRes}
\alias{getOptionsFromRes}
\title{A function to derive the option details from the resources data frame returned in the function above, specifically with the ohlc underlying attribute.}
\usage{
getOptionsFromRes(session, date = Sys.Date())
}
\arguments{
\item{session}{list of the rdecaf session.}

\item{date}{date of the as-of/up-to date to query active and unexpired instruments. Defaults to current day.}
}
\value{
The data-frame with identified underlying instruments of options and relevant info, including account details.
}
\description{
This is a description.
}
