% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdings.R
\name{getOrderedHoldings}
\alias{getOrderedHoldings}
\title{A function to generate a ordered holdings data-frame}
\usage{
getOrderedHoldings(holdings, toplevel = "Subtype", sublevels = c("CCY",
  "Country"), customTop = NULL, ...)
}
\arguments{
\item{holdings}{The ordered holdings.}

\item{toplevel}{TODO}

\item{sublevels}{TODO}

\item{customTop}{TODO}

\item{...}{Additional parameters}
}
\value{
Returns the nested holdings data-frame.
}
\description{
This is the description
}
