% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{getPconsolidation}
\alias{getPconsolidation}
\title{This function gets the pconsolidations for a portfolio.}
\usage{
getPconsolidation(portfolio, session)
}
\arguments{
\item{portfolio}{The portfolio id.}

\item{session}{The rdecaf session.}
}
\value{
A list with the pconsolidations for the portfolio.
}
\description{
This is the description
}
