% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{getPipeline}
\alias{getPipeline}
\title{A function to create the file pipeline for decaf-imports.}
\usage{
getPipeline(isLocal, providers = NULL, append = NULL,
  processedFiles = FALSE, ignores = c(".zip", ".xml", ".txt",
  "provider_resource_id.csv"))
}
\arguments{
\item{isLocal}{Boolean to indicate whether we are in a test environment. Default is FALSE.}

\item{providers}{A vector with file provider names. If NULL, all existing providers will be considers. Default is NULL.}

\item{append}{A vector with non-file data providers (i.e API). Any non-file data provider will be appended to the pipeline. Default is NULL.}

\item{processedFiles}{Boolean to indicate whether processed files should be included. Default is FALSE.}

\item{ignores}{A vector with key words to ignore for the pipeline.}
}
\value{
A reference string.
}
\description{
This is the description
}
