% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{getPJournalsByAccount}
\alias{getPJournalsByAccount}
\title{This function gets the partial journal entries for a specific account id.}
\usage{
getPJournalsByAccount(id, session)
}
\arguments{
\item{id}{The account id.}

\item{session}{The rdecaf session.}
}
\value{
A data frame with the partial journal entries.
}
\description{
This is the description
}
