% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnl.R
\name{getPnlPreemble}
\alias{getPnlPreemble}
\title{A function to provide the pnl preemble.}
\usage{
getPnlPreemble(
  posBeg,
  posEnd,
  resources,
  quants,
  trades,
  portfolio,
  fxobs,
  ccy,
  session
)
}
\arguments{
\item{posBeg}{The beginning period positions from getFlatHoldings.}

\item{posEnd}{The ending period positions from getFlatHoldings.}

\item{resources}{The resources from rdecaf.}

\item{quants}{The quants data frame.}

\item{trades}{The trades data frame.}

\item{portfolio}{The portfolio id.}

\item{fxobs}{The fx rates.}

\item{ccy}{The portfolio currency.}

\item{session}{The rdecaf session.}
}
\value{
A list with the pnl preemble.
}
\description{
This is the description
}
