% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{getPortfolio}
\alias{getPortfolio}
\title{This function gets the portfolio information give a portfolio id.}
\usage{
getPortfolio(portfolio, session)
}
\arguments{
\item{portfolio}{The portfolio id.}

\item{session}{The rdecaf session.}
}
\value{
A data frame with the portfolio details.
}
\description{
This is the description
}
