% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{getPortfolioWeights}
\alias{getPortfolioWeights}
\title{A function to compute the return the portfolio weights over a specific periodicity.}
\usage{
getPortfolioWeights(
  p,
  session,
  start = Sys.Date() - 360,
  end = Sys.Date(),
  period = "W"
)
}
\arguments{
\item{p}{The portfolio id.}

\item{session}{The session of the instance.}

\item{start}{The start date. Default is 360 days ago from today.}

\item{end}{The end date. Default is today.}

\item{period}{The memnonic for dateOfPeriod function. Any of D, W, M, Q, Y.}
}
\value{
A list with the input data and the results, where results are the weights over the periods
}
\description{
A function to compute the return the portfolio weights over a specific periodicity.
}
