% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{getPortNameByAccount}
\alias{getPortNameByAccount}
\title{A function to map portfolio info to a trades data frame by account id, referenced as "account".}
\usage{
getPortNameByAccount(data, session)
}
\arguments{
\item{data}{a trades data frame}

\item{session}{The rdecaf session}
}
\value{
A data frame
}
\description{
This is the description
}
