% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionIncrement.R
\name{getPositionComparisonNEW}
\alias{getPositionComparisonNEW}
\title{This function masks decaf positions based on provider positions:}
\usage{
getPositionComparisonNEW(accPos, resources, posDate, group, useCostPX, session)
}
\arguments{
\item{accPos}{The positions data frame from the provider.}

\item{resources}{The resources data frame.}

\item{posDate}{The position date.}

\item{group}{A vector with the desired resource ctypes.}

\item{useCostPX}{For empty consolidations, shall cost price be used?}

\item{session}{The rdecaf session.}
}
\value{
Returns the side by side comparison of decaf position and provider position.
}
\description{
This is the description
}
