% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{getPositionsByGrp}
\alias{getPositionsByGrp}
\title{Provides the positions by groupings.}
\usage{
getPositionsByGrp(holdings, col)
}
\arguments{
\item{holdings}{The holdings data frame.}

\item{col}{The columns to be arranged by.}
}
\value{
A list.
}
\description{
This is a description.
}
