% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdings.R
\name{getPrintableHoldings}
\alias{getPrintableHoldings}
\title{A function to generate the printable holdings data-frame}
\usage{
getPrintableHoldings(portfolio, ccy, date, dtype, toplevel, sublevel,
  customTop, colselect, regions, summaryaddon, addTagsBy = NULL,
  childDefaults = TRUE, charLimit = 30, orderByKeys = c("Expiry",
  "CCY", "Exposure", "Name"), session, ...)
}
\arguments{
\item{portfolio}{The portfolio id}

\item{ccy}{The currency of reporting}

\item{date}{The date of reporting}

\item{dtype}{The date type of reporting}

\item{toplevel}{TODO}

\item{sublevel}{TODO}

\item{customTop}{TODO}

\item{colselect}{TODO}

\item{regions}{The list with the country to region mapping}

\item{summaryaddon}{TODO}

\item{addTagsBy}{TODO}

\item{childDefaults}{TODO}

\item{charLimit}{TODO}

\item{orderByKeys}{TODO}

\item{session}{The rdecaf session}

\item{...}{Additional parameters}
}
\value{
A data-frame with the printable format of holdings
}
\description{
This is a description.
}
