% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{getPXLast}
\alias{getPXLast}
\title{This function mimicks the PX Last methodology of DECAF:}
\usage{
getPXLast(resmain, resources, session, date)
}
\arguments{
\item{resmain}{The resmain id in DECAF:}

\item{resources}{The rdecaf resources:}

\item{session}{The rdecaf session}

\item{date}{The date of interest.}
}
\value{
Returns the last available price for the resmains
}
\description{
This is the description
}
