% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRandString}
\alias{getRandString}
\title{A function to produce a random string.}
\usage{
getRandString(n = 1, len = 12)
}
\arguments{
\item{n}{The number of random strings.}

\item{len}{The desired lenght of the random string.}
}
\value{
A random string.
}
\description{
A function to produce a random string.
}
