% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocketBase.R
\name{getRecs}
\alias{getRecs}
\title{Pocketbase dependent function to gather records from the db.}
\usage{
getRecs(rb, path, ext, item)
}
\arguments{
\item{rb}{The rocketbase authorizaion credentials. Defined externally.}

\item{path}{The path of the pocketbase API endpoint, e.g. /api/collections/cfraEquities/records}

\item{ext}{the ext to the path - where applicable, "?page=1"}

\item{item}{The actual endpoint element to return, e.g. 'items' or 'pages'}
}
\value{
a data frame or vector of the item(s).
}
\description{
This is the description
}
