% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remapDB.R
\name{getRemapContent}
\alias{getRemapContent}
\title{A function to get the remap content on strapi.}
\usage{
getRemapContent(baseUrl = "https://db.remap.decafhub.com/", contentName)
}
\arguments{
\item{baseUrl}{The base url of the remap db. Default: "https://db.remap.decafhub.com/".}

\item{contentName}{The name of the content.}
}
\value{
The httr GET response.
}
\description{
This is a description.
}
