% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbg.R
\name{getRequestDataPrice}
\alias{getRequestDataPrice}
\title{A function to build the requestable currency pairs and the entire combinations of currencies}
\usage{
getRequestDataPrice(session, date, zero, underlying, fx = NULL,
  resources, priceTag, byStock)
}
\arguments{
\item{session}{The rdecaf session.}

\item{date}{The date.}

\item{zero}{The zero parameter for stocks.}

\item{underlying}{Shall underlying be included if stocks are to be considered?}

\item{fx}{If exchange rate prices are to be called, provide a vector with main currencies, otherwise NULL.}

\item{resources}{The entire resource data frame.}

\item{priceTag}{Force resources which this tag to be considered.}

\item{byStock}{Should the request data be based on stocks only?}
}
\value{
A list with the reference request data objest and the resources.
}
\description{
A function to build the requestable currency pairs and the entire combinations of currencies
}
