% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{getResourcesByStock}
\alias{getResourcesByStock}
\title{A function to retrieve the resources by stock for a rdecaf session:}
\usage{
getResourcesByStock(stocks, session)
}
\arguments{
\item{stocks}{The stocks data-frame}

\item{session}{The rdecaf session}
}
\value{
Returns the inferred maturity
}
\description{
This is a description.
}
