% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{getResUndrAttr}
\alias{getResUndrAttr}
\title{A function to query decaf resources endpoint for a specific attribute.}
\usage{
getResUndrAttr(
  session,
  type = "OPT",
  date = Sys.Date(),
  attr = "attributes.ohlc_underlying"
)
}
\arguments{
\item{session}{list of the rdecaf session.}

\item{type}{string of the ctype to filter for, defaults to options.}

\item{date}{date of the as-of/up-to date to query active and unexpired instruments. Defaults to current day.}

\item{attr}{string of the attributes column identifying the underlying. Defaults to ohlc_underlying.}
}
\value{
The resources data-frame filtered.
}
\description{
This is a description.
}
