% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{getRf}
\alias{getRf}
\title{Gets the riskfree elements for a portfolio.}
\usage{
getRf(rfSymbol = NULL, start, end, session)
}
\arguments{
\item{rfSymbol}{the risk free rate, denoted by a named (code or value) list containing the value (constant decimal or ohlc series id).}

\item{start}{The desired start date.}

\item{end}{The desired end date.}

\item{session}{The rdecaf session.}
}
\value{
a time series containing risk free rate.
}
\description{
This is a description.
}
