% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{getStocks}
\alias{getStocks}
\title{A function to retrieve the stocks for set of accounts:}
\usage{
getStocks(container, session, zero = 1, date = Sys.Date(), c = "account")
}
\arguments{
\item{container}{The container data-frame from rdecaf}

\item{session}{The rdecaf session}

\item{zero}{Either 1 or 0 to indicate whether closed positions should be included.}

\item{date}{The as of date for the stocks. Default is today.}

\item{c}{The type. Either "account" or "portfolio".}
}
\value{
Returns a data-frame with the stocks for the accounts.
}
\description{
This is a description.
}
