% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnl.R
\name{getStocksAndAssets}
\alias{getStocksAndAssets}
\title{flattens the pnl endpoint list data from decaf.}
\usage{
getStocksAndAssets(session, portfolio = NULL, date = Sys.Date())
}
\arguments{
\item{session}{the decaf session.}

\item{portfolio}{the decaf portfolio ID. Defaults to NULL.}

\item{date}{the as of date. Defaults to present.}
}
\value{
A flat df containing the session artifacts and associated asset class.
}
\description{
This is the description
}
