% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{getStocksFromContainerNames}
\alias{getStocksFromContainerNames}
\title{A function to get stocks from container type and names.}
\usage{
getStocksFromContainerNames(
  session,
  containerType,
  containerNames,
  zero = 1,
  date
)
}
\arguments{
\item{session}{The rdecaf session:}

\item{containerType}{The type of container, either 'portfolios' or 'accounts'.}

\item{containerNames}{The names of the container.}

\item{zero}{If 1, closed and open stocks are considered. If 0, only open stocks.}

\item{date}{The date of the stocks.}
}
\value{
Returns the data frame with the stocks.
}
\description{
This is a description.
}
