% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getStringDistanceIndex}
\alias{getStringDistanceIndex}
\title{This function}
\usage{
getStringDistanceIndex(
  v1,
  v2,
  param = 0.2,
  dir = "smaller",
  method = "jaccard"
)
}
\arguments{
\item{v1}{A vector with strings to be matched.}

\item{v2}{A vector with candidate matching strings.}

\item{param}{The stringdist method's parameter.}

\item{dir}{Bigger or smaller than the param?}

\item{method}{The method for stringdist.}
}
\value{
A ordered vector
}
\description{
This is a description.
}
