% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{getTradeByCtypeForPortfolio}
\alias{getTradeByCtypeForPortfolio}
\title{A function to sync trades in batches}
\usage{
getTradeByCtypeForPortfolio(ctype, portfolio, gte = NULL, session)
}
\arguments{
\item{ctype}{The trade ctype.}

\item{portfolio}{The portfolio id.}

\item{gte}{Great than date.}

\item{session}{The rdecaf session.}
}
\value{
A data frame with the trades for desired ctype.
}
\description{
This is the description
}
