% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{getTradeIdentity}
\alias{getTradeIdentity}
\title{For a incoming trade record, this function identifies corresponding candidate trades in decaf.}
\usage{
getTradeIdentity(records, accmain, session)
}
\arguments{
\item{records}{A data frame with incoming trade records. Requires columns 'resmain', 'commitment' and 'settlement'}

\item{accmain}{The account id in decaf}

\item{session}{The rdecaf session info}
}
\value{
A list of length NROW(records).
}
\description{
This is the description
}
