% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{getTradesFromContainerNames}
\alias{getTradesFromContainerNames}
\title{A function to get trades from session using container names:}
\usage{
getTradesFromContainerNames(containerNames, session, type, gte = NULL)
}
\arguments{
\item{containerNames}{A vector with container names}

\item{session}{The DECAF session info}

\item{type}{The container type}

\item{gte}{A date object to express 'Greater Than' for commitment date of trades. Default is NULL.}
}
\value{
A data-frame with DECAF trades for container.
}
\description{
This is the description
}
