% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{getTradesFromPortfolioNames}
\alias{getTradesFromPortfolioNames}
\title{A function to get trades from session using portfolio names:}
\usage{
getTradesFromPortfolioNames(portfolioNames, session)
}
\arguments{
\item{portfolioNames}{A vector with portfolio names:}

\item{session}{The DECAF session info.}
}
\value{
A data-frame with DECAF trades for portfolio.
}
\description{
This is the description
}
