% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeorder.R
\name{getTradingObjects}
\alias{getTradingObjects}
\title{A function to retrieve trade order data objects.}
\usage{
getTradingObjects(session)
}
\arguments{
\item{session}{The rdecaf session.}
}
\value{
A list with trader oder, trade order allocation and trade order execution data frames.
}
\description{
This is the description
}
