% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnl.R, R/trades.R
\name{getTransferValueAmounts}
\alias{getTransferValueAmounts}
\title{Gets the portfolio level amounts tagged as transfer.}
\usage{
getTransferValueAmounts(portfolio, since, until, currency, session)

getTransferValueAmounts(portfolio, since, until, currency, session)
}
\arguments{
\item{portfolio}{The portfolio id.}

\item{since}{The date since.}

\item{until}{The date until.}

\item{currency}{The portfolio currency.}

\item{session}{The rdecaf session.}
}
\value{
data frame containing transfer values by trade date.

A data frame with the converted value amounts of transfers.
}
\description{
This is the description

This is the description
}
