% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{getTXNsOfTrades}
\alias{getTXNsOfTrades}
\title{A function to sync trades in batches}
\usage{
getTXNsOfTrades(trades, ccy = NULL, session)
}
\arguments{
\item{trades}{The trades data frame.}

\item{ccy}{The currency to convert valamt's to.}

\item{session}{The rdecaf session.}
}
\value{
A data frame with the trades for desired ctype.
}
\description{
This is the description
}
