% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{getUnderlyingOptions}
\alias{getUnderlyingOptions}
\title{A function to compare px on option contract strike and its underlying security price to alert users beyond a threshold.}
\usage{
getUnderlyingOptions(
  session,
  type = "OPT",
  date = Sys.Date(),
  attr = "attributes.ohlc_underlying",
  thresh = 0.1
)
}
\arguments{
\item{session}{list of the rdecaf session.}

\item{type}{string of the ctype to filter for, defaults to options.}

\item{date}{date of the as-of/up-to date to query active and unexpired instruments. Defaults to current day.}

\item{attr}{string of the attributes column identifying the underlying. Defaults to ohlc_underlying.}

\item{thresh}{numeric threshold setting the limit of the px convergence to trigger an email alert for. Defaults to 10%.}
}
\value{
The data-frame with identified underlying instruments of options and relevant info.
}
\description{
This is a description.
}
