% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{getUnderlyingSymbol}
\alias{getUnderlyingSymbol}
\title{A function to get underlying symbol of an instrument.}
\usage{
getUnderlyingSymbol(resources, instring = "Equity", sweep = FALSE)
}
\arguments{
\item{resources}{data frame containing the resources endpoint}

\item{instring}{string dictating what instrument type to search the underlying symbol for. Defaults to Equity.}

\item{sweep}{logical to dictate whether to overwrite values with existing values or not. Defaults to false.}
}
\value{
data frame with the cleaned up underlying column.
}
\description{
This is the description
}
