% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{getUsageMetrics}
\alias{getUsageMetrics}
\title{This function to get the usage metrics.}
\usage{
getUsageMetrics(userActions, tokenSession)
}
\arguments{
\item{userActions}{The user actions result.}

\item{tokenSession}{The token session data frame.}
}
\value{
A xts usage metrics data frame.
}
\description{
This function to get the usage metrics.
}
