% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getWTDSlice}
\alias{getWTDSlice}
\title{A function which provides the WTD slice of the data frame.}
\usage{
getWTDSlice(df, date)
}
\arguments{
\item{df}{The data frame with a 'date' column.}

\item{date}{the report date.}
}
\value{
A data frame with the WTD slice of the data frame or NULL.
}
\description{
This is the description
}
