% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav.R
\name{htmlSummary}
\alias{htmlSummary}
\title{Styling function that creates links/pretty numbers to be shared, given summarized outlier data and deployment}
\usage{
htmlSummary(df, deployment)
}
\arguments{
\item{df}{the summarized outlier data frame from above or customized. Requires date and portfolio (decaf ID) as columns along with others in returned summary data frame from outlierwrapper.}

\item{deployment}{the client name as found in respective URLs}
}
\value{
A dataframe with the stylized summary output.
}
\description{
This is the description
}
