% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{inbulkAccount}
\alias{inbulkAccount}
\title{A function to inbulk accounts.}
\usage{
inbulkAccount(
  guid = NULL,
  name,
  portfolio,
  custodian,
  guidPrefix = NULL,
  xguid = NULL,
  session
)
}
\arguments{
\item{guid}{The vector with guids. If NULL (default), creates using guidPrefix and xguid.}

\item{name}{The vector with names.}

\item{portfolio}{The vector with portfolio id's.}

\item{custodian}{The vector with custodian id's.}

\item{guidPrefix}{A string to use as prefix (e.g Custodian Name)}

\item{xguid}{A vector with the id to be transformed to guid.}

\item{session}{The rdecaf session}
}
\value{
Returns NULL
}
\description{
This is the description
}
