% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{inbulkInvestment}
\alias{inbulkInvestment}
\title{A function to inbulk accounts.}
\usage{
inbulkInvestment(
  guid = NULL,
  resmain,
  accmain,
  agent,
  commitment,
  pxnavs,
  shrcnt,
  qtymain,
  shrcls,
  notes = NA,
  session
)
}
\arguments{
\item{guid}{The vector with guids. If NULL (default), created inside.}

\item{resmain}{TODO.}

\item{accmain}{TODO.}

\item{agent}{TODO.}

\item{commitment}{TODO.}

\item{pxnavs}{TODO.}

\item{shrcnt}{TODO.}

\item{qtymain}{TODO.}

\item{shrcls}{TODO.}

\item{notes}{TODO.}

\item{session}{The rdecaf session}
}
\value{
A list with results.
}
\description{
This is the description
}
