% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{inbulkPortfolio}
\alias{inbulkPortfolio}
\title{A function to inbulk portfolios.}
\usage{
inbulkPortfolio(
  guid = NULL,
  name,
  rccy,
  team,
  guidPrefix = NULL,
  xguid = NULL,
  session
)
}
\arguments{
\item{guid}{The vector with guids. If NULL (default), creates using guidPrefix and xguid.}

\item{name}{The vector with names.}

\item{rccy}{The vector of currencies.}

\item{team}{The vector of teams id's.}

\item{guidPrefix}{A string to use as prefix (e.g Custodian Name)}

\item{xguid}{A vector with the id to be transformed to guid.}

\item{session}{The rdecaf session}
}
\value{
Returns NULL
}
\description{
This is the description
}
