% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{inbulkTrades}
\alias{inbulkTrades}
\title{A function to inbulk trades.}
\usage{
inbulkTrades(data, guid = NULL, guidPrefix = NULL, xguid = NULL,
  session)
}
\arguments{
\item{data}{The data frame with trade records.}

\item{guid}{The vector with guids. If NULL (default), creates using guidPrefix and xguid.}

\item{guidPrefix}{A string to use as prefix (e.g Custodian Name)}

\item{xguid}{A vector with the id to be transformed to guid.}

\item{session}{The rdecaf session}
}
\value{
Returns list with inbulk results.
}
\description{
This is the description
}
