% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocketBase.R
\name{incrIt}
\alias{incrIt}
\title{Pocketbase dependent function to push data efficiently without duplicating already existing data in the DB.}
\usage{
incrIt(rb, path, increment, join, push = TRUE)
}
\arguments{
\item{rb}{The rocketbase authorizaion credentials. Defined externally.}

\item{path}{The path of the pocketbase API endpoint, e.g. /api/collections/cfraEquities/records}

\item{increment}{The data to add or append to the endpoint.}

\item{join}{The vector of the keys to join by to exclude already existing data in the DB.}

\item{push}{The TRUE/FALSE flag dictating whether to push the data or just return the records.}
}
\value{
NULL or the records depending on the outcome.
}
\description{
This is the description
}
