% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{indexCompiler}
\alias{indexCompiler}
\title{TODO:}
\usage{
indexCompiler(df, inCols, exCols, idCols, addConditions)
}
\arguments{
\item{df}{A data-frame.}

\item{inCols}{TODO}

\item{exCols}{TODO}

\item{idCols}{TODO}

\item{addConditions}{TODO}
}
\value{
A list with the row indices and the group indices of the strategy.
}
\description{
TODO:
}
