% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{inferReversals}
\alias{inferReversals}
\title{## A function to infer reversal trades}
\usage{
inferReversals(trades)
}
\arguments{
\item{trades}{The trades data frame with pxmain, accmain, commitment and qytmain columns:}
}
\value{
The same data frame with reversal column added
}
\description{
This is the description
}
