% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{inferReversalsV2}
\alias{inferReversalsV2}
\title{## A function to infer reversal trades}
\usage{
inferReversalsV2(trades)
}
\arguments{
\item{trades}{The trades data frame with pxmain, accmain, commitment, qytmain, and resmain columns:}
}
\value{
The same data frame with reversal column added
}
\description{
This is the description
}
