% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initDF}
\alias{initDF}
\title{A function initialises a data frame with desired colnames and rows.}
\usage{
initDF(colNames, nRow = 1)
}
\arguments{
\item{colNames}{A vector with desired column names.}

\item{nRow}{The desired number of rows. Default is 1.}
}
\value{
A data-frame with NA's
}
\description{
This is a description.
}
