% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remapDB.R
\name{isBBGTicker}
\alias{isBBGTicker}
\title{This function tries to identify if string is a bloomberg ticker.}
\usage{
isBBGTicker(tickers)
}
\arguments{
\item{tickers}{A vector with strings.}
}
\value{
Returns a TRUE | FALSE vector.
}
\description{
This is a description.
}
