% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{isinCCYMatch}
\alias{isinCCYMatch}
\title{A function to match instrument in a data frame with resources in the system.}
\usage{
isinCCYMatch(data, resources)
}
\arguments{
\item{data}{A data-frame with 'isin' and 'ccymain' columns}

\item{resources}{The resources data-frame from rdecaf}
}
\value{
The data-frame with the resmain column
}
\description{
This is a description.
}
