% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isIsin}
\alias{isIsin}
\title{A function to determine whether string resembles ISIN:}
\usage{
isIsin(str)
}
\arguments{
\item{str}{A vector with strings.}
}
\value{
A vector with either string or NA.
}
\description{
This is a description.
}
