% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isNAorEmpty}
\alias{isNAorEmpty}
\title{A function to check whether value is empty or NA:}
\usage{
isNAorEmpty(str)
}
\arguments{
\item{str}{A vector of strings.}
}
\value{
TRUE or FALSE.
}
\description{
This is a description.
}
