% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remapDB.R
\name{isTelekursID}
\alias{isTelekursID}
\title{This function tries to identify if string is a telekurs id.}
\usage{
isTelekursID(str)
}
\arguments{
\item{str}{A vector with strings.}
}
\value{
Returns a TRUE | FALSE vector.
}
\description{
This is a description.
}
