% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isWholenumber}
\alias{isWholenumber}
\title{This function checks if value is a whole number:}
\usage{
isWholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{A vector with values.}

\item{tol}{The tolerance.}
}
\value{
Returns a boolean vector.
}
\description{
This is a description.
}
