% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juliusbaerCommons.R
\name{juliusbaerPosColumnMappers}
\alias{juliusbaerPosColumnMappers}
\title{A function to return the column mappers}
\usage{
juliusbaerPosColumnMappers()
}
\value{
Returns a list with the column mappers.
}
\description{
This is the description
}
