% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juliusbaerCommons.R
\name{juliusbaerSecTypeMapper}
\alias{juliusbaerSecTypeMapper}
\title{A function to return the security type mapper}
\usage{
juliusbaerSecTypeMapper()
}
\value{
Returns a list with the security type mappers.
}
\description{
This is the description
}
