% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juliusbaerCommons.R
\name{juliusbaerTraColumnMappers}
\alias{juliusbaerTraColumnMappers}
\title{A function to return the column mappers}
\usage{
juliusbaerTraColumnMappers()
}
\value{
Returns a list with the column mappers.
}
\description{
This is the description
}
