% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{keyStats}
\alias{keyStats}
\title{Provides the keystatistics given the input of the getPerformanceV2 output and an as-of-date}
\usage{
keyStats(performance, date)
}
\arguments{
\item{performance}{list of output from function referenced above.}

\item{date}{the asofdate.}
}
\value{
A list with return statistics.
}
\description{
This is the description
}
