% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lombardCommons.R
\name{lombardPosColumnMappers}
\alias{lombardPosColumnMappers}
\title{A function to return the column mappers}
\usage{
lombardPosColumnMappers()
}
\value{
Returns a list with the column mappers.
}
\description{
This is the description
}
