% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lombardCommons.R
\name{lombardSecTypeMapper}
\alias{lombardSecTypeMapper}
\title{A function to return the security type mapper}
\usage{
lombardSecTypeMapper()
}
\value{
Returns a list with the security type mappers.
}
\description{
This is the description
}
