% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{makeImgTagSVG}
\alias{makeImgTagSVG}
\title{Creates an image tag for SVG.}
\usage{
makeImgTagSVG(path)
}
\arguments{
\item{path}{The path to the SVG image.}
}
\value{
An image tag.
}
\description{
This is a description.
}
