% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionIncrement.R
\name{maskPositions}
\alias{maskPositions}
\title{This function masks decaf positions based on provider positions:}
\usage{
maskPositions(pPos, resources, pDate)
}
\arguments{
\item{pPos}{The positions data frame from the provider.}

\item{resources}{The resources data frame}

\item{pDate}{The position date.}
}
\value{
Returns a masked decaf positions data frame.
}
\description{
This is the description
}
