% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{matchFJE}
\alias{matchFJE}
\title{A function returns the full journal entry in the system which matches the record most.}
\usage{
matchFJE(qtymain, commitment, portfolio, keyword = NA, session)
}
\arguments{
\item{qtymain}{The quantity of the record.}

\item{commitment}{The date of the record.}

\item{portfolio}{The portfolio to be considered.}

\item{keyword}{Optional. The keyword in FJE notes to be considered.}

\item{session}{The rdecaf session}
}
\value{
A list with results.
}
\description{
This is the description
}
