% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{matchTradeIdentity}
\alias{matchTradeIdentity}
\title{For trade identity candidates, infer matches}
\usage{
matchTradeIdentity(tradeIdentities)
}
\arguments{
\item{tradeIdentities}{A list with the trade identity candidates.}
}
\value{
A list with matched trade identities.
}
\description{
This is the description
}
