% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mCondition}
\alias{mCondition}
\title{In a data frame, check if multiple columns fulfil multiple conditions.}
\usage{
mCondition(df, cols, conditions)
}
\arguments{
\item{df}{The data frame.}

\item{cols}{A vectors of column names.}

\item{conditions}{The conditions against which to check the columns.}
}
\value{
A vector with TRUE or FALSE.
}
\description{
This is the description
}
