% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mgsub}
\alias{mgsub}
\title{A function to gsub multiple patterns.}
\usage{
mgsub(str, patterns, replace = "")
}
\arguments{
\item{str}{The original string.}

\item{patterns}{A vector with string patterns.}

\item{replace}{A string to replace the patterns with. Default is "".}
}
\value{
The new string.
}
\description{
A function to gsub multiple patterns.
}
