% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mmatch}
\alias{mmatch}
\title{A function to match a vector with a data frame with multiple columns.}
\usage{
mmatch(matchFrom, matchTo)
}
\arguments{
\item{matchFrom}{The vector to be matched.}

\item{matchTo}{The data frame to be matched to.}
}
\value{
A vector with the matching indices.
}
\description{
This is the description
}
