% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketStack.R
\name{mstackSrchTicker}
\alias{mstackSrchTicker}
\title{Returns df of tickers from marketstack}
\usage{
mstackSrchTicker(
  access,
  search = "",
  searchOrExchange = "&exchange=",
  lim = 1000,
  EOD = c(TRUE),
  country = "US"
)
}
\arguments{
\item{access}{marketstack API key}

\item{search}{free text value to search for e.g. ETF, ARCX (NYSE). Defaults to empty}

\item{searchOrExchange}{are we filtering my name/ticker or exchange, default is exchange, change to &search for ticker}

\item{lim}{limit of records to return, defaults to max of 1000}

\item{EOD}{filter to look at just data that EOD, defaults to true}

\item{country}{if we want to filter by country 2-digit code for currency. Defaults to US}
}
\value{
A data frame with market data.
}
\description{
This is a description.
}
