% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{mTSPlot}
\alias{mTSPlot}
\title{Provides a line time series plot, with series for optional benchmarks.}
\usage{
mTSPlot(
  df,
  date,
  smooth = 0.1,
  limitFactor = 0.05,
  title,
  col = "#79CDCD",
  lout = 10,
  palette_all = NULL,
  labels = NULL,
  legpos = "topright",
  ylab = "Index",
  grid = 1,
  ...
)
}
\arguments{
\item{df}{A data frame with columns of all time series values, col1 is the target, 2:n are the benchmarks.}

\item{date}{A date vector (needs to be same length as the timeseries df) of x axis date values.}

\item{smooth}{The smoothing parameter.}

\item{limitFactor}{The factor with which to expand the limits.}

\item{title}{The title of the plot.}

\item{col}{The color for the polygon.}

\item{lout}{The number of ticks to display.}

\item{palette_all}{An optional vector parameter of colors to assign to addl lines in same order as df e.g. element 1 is the first benchmark color.}

\item{labels}{An optional vector parameter of legend labels for target and n benchmarks - needs to be same length as df.}

\item{legpos}{The legend position.}

\item{ylab}{The y axis header.}

\item{grid}{A parameter to decide whether to add line grids background.}

\item{...}{Option for additional params.}
}
\value{
An line time series plot.
}
\description{
This is the description
}
