% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{multicol_spec}
\alias{multicol_spec}
\title{A function to style multiple cols in data frame}
\usage{
multicol_spec(x, cols, ...)
}
\arguments{
\item{x}{The data frame.}

\item{cols}{The column indices.}

\item{...}{Arguments for kablExtra.}
}
\value{
A data frame with style rows.
}
\description{
This is a description.
}
