% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{multirow_spec}
\alias{multirow_spec}
\title{A function to style multiple rows in data frame}
\usage{
multirow_spec(x, rows, ...)
}
\arguments{
\item{x}{The data frame.}

\item{rows}{The row indices.}

\item{...}{Arguments for kablExtra.}
}
\value{
A data frame with style rows.
}
\description{
This is a description.
}
