% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mxgrep}
\alias{mxgrep}
\title{This function greps single key in all columns of a data-frame.}
\usage{
mxgrep(df, key)
}
\arguments{
\item{df}{A data-frame.}

\item{key}{The key to be grepped.}
}
\value{
Returns a data-frame with the grepped key.
}
\description{
This is a description.
}
