% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nearesth}
\alias{nearesth}
\title{A function to round up/down to the the nearest value.}
\usage{
nearesth(x, nearest = 0.1, dir = "up")
}
\arguments{
\item{x}{An array of numeric values.}

\item{nearest}{The value to be rounded to.}

\item{dir}{The direction to be rounded to (up or down).}
}
\value{
Formated number(s):
}
\description{
A function to round up/down to the the nearest value.
}
