% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nestedOrdering}
\alias{nestedOrdering}
\title{A function to order a data-frame based on multiple columns:}
\usage{
nestedOrdering(df, colNames)
}
\arguments{
\item{df}{A data-frame.}

\item{colNames}{The names of the columns to be considered for order. The columns must be numeric.}
}
\value{
The ordered data-frame.
}
\description{
A function to order a data-frame based on multiple columns:
}
