% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numberEquivalency}
\alias{numberEquivalency}
\title{This function checks if 2 numbers match}
\usage{
numberEquivalency(num1, num2, pctTolerance = 0.001, absTolerance = 0.05)
}
\arguments{
\item{num1}{The first number}

\item{num2}{The second number}

\item{pctTolerance}{The percentage tolerance}

\item{absTolerance}{The absolute tolerance}
}
\value{
TRUE or FALSEg
}
\description{
This is a description.
}
