% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numerize}
\alias{numerize}
\title{This function attempts to transform irregular number
formats into numerical values.}
\usage{
numerize(str)
}
\arguments{
\item{str}{A vector of strings.}
}
\value{
A vector of numerical values.
}
\description{
This is a description.
}
