% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health.R
\name{ohlcHealthWrapper}
\alias{ohlcHealthWrapper}
\title{This is a wrapper function for ohlc health function.}
\usage{
ohlcHealthWrapper(
  asof = Sys.Date(),
  ohlccodes = NULL,
  underlying = FALSE,
  lookBack = 365,
  session
)
}
\arguments{
\item{asof}{The asof date. Default is Sys.Date().}

\item{ohlccodes}{The ohlccodes to be check. If NULL, it gets it from the stocks. Default is NULL.}

\item{underlying}{Shall the underlying be checked as well? Default is FALSE}

\item{lookBack}{The lookback periodin days. Default is 365.}

\item{session}{The rdecaf session.}
}
\value{
A data frame with unhealthy ohlc observations.
}
\description{
This is a description
}
