% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{omitRecordsByFlag}
\alias{omitRecordsByFlag}
\title{When syncing between 2 decaf instances, this function helps to omit syncing of records by cflag.}
\usage{
omitRecordsByFlag(endpoint, sourceData, session, omitFlag, addParams = NULL)
}
\arguments{
\item{endpoint}{Which decaf endpoint is this regarding? i.e 'resources'.}

\item{sourceData}{The data-frame of the source of the same endpoint. i.e the 'resources' data frame at source.}

\item{session}{The target session}

\item{omitFlag}{The cflag by which we shall omit? i.e 1.}

\item{addParams}{Additional parameters to be added to getDBObjects(endpoint). Default=NULL.}
}
\value{
The source data frame without the cflagged recrods at target.
}
\description{
This is the description
}
