% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{optionDelta}
\alias{optionDelta}
\title{A function to produce the option deltas.}
\usage{
optionDelta(holdings, resources)
}
\arguments{
\item{holdings}{The holdings.}

\item{resources}{The resources.}
}
\value{
The option deltas.
}
\description{
This is the description
}
