% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{optionsAlert}
\alias{optionsAlert}
\title{A function to taking a list of emails and portfolios, creating an options alert with content derived from given data.}
\usage{
optionsAlert(
  data,
  portfolios = NULL,
  title = "Options Alert",
  from = "info@telostat.com",
  emails = list("andre@telostat.com"),
  thresh = 0.1
)
}
\arguments{
\item{data}{dataframe, such as the output from function above, containing at a minimum porfolio_name and delta columns.}

\item{portfolios}{list of portfolio names to filter for in the data, where applicable. Defaults to null.}

\item{title}{string for the message body of the email alert. Defaulst to options alert.}

\item{from}{string for the sender email address. Defaults to telostat.}

\item{emails}{list of email strings for the recipient email addresses. Defaults to andre.}

\item{thresh}{numeric threshold setting the limit of the px convergence filter for delta to trigger an email alert for. Defaults to 0.1.}
}
\value{
null with a printed message confirming send.
}
\description{
This is a description.
}
