% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orderByKey}
\alias{orderByKey}
\title{This function orders a vector by the keys provided.}
\usage{
orderByKey(x, keys)
}
\arguments{
\item{x}{A vector}

\item{keys}{The vector of keys in order of preference.}
}
\value{
A ordered vector
}
\description{
This is a description.
}
