% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav.R
\name{outlierData}
\alias{outlierData}
\title{Aggregating function that runs the outlier wrapper for an entire sessions portfolios, given a lookback period}
\usage{
outlierData(session, asOf = Sys.Date(), lookBack = 90, resources = NULL)
}
\arguments{
\item{session}{the rdecaf session.}

\item{asOf}{the as of date for the lookback. Defaults to present.}

\item{lookBack}{the horizon of the lookback. Defaults to 90 days.}

\item{resources}{the decaf endpoint if provided. Defaults to NULL and is queried.}
}
\value{
A list with 2 data frame elements, the first containing the portfolio level summary, the second the date-level details.
}
\description{
This is the description
}
