% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav.R
\name{outlierWrapper}
\alias{outlierWrapper}
\title{Wrapper function that runs find and clean outlier functions above for a given portfolio or set of portfolios}
\usage{
outlierWrapper(
  date = Sys.Date(),
  resources,
  excluded = c("FUT", "FXFWD"),
  inferred = c("LOAN"),
  external = FALSE,
  tradeLvl = FALSE,
  params = c(correct = 0.5, thresh = 0.02, toler = 0.2, ceiling = 8, floor = 3.5),
  portfolios,
  session,
  lookBack = 365
)
}
\arguments{
\item{date}{the end date for the analysis. Defaults to present.}

\item{resources}{the rdecaf resources endpoint.}

\item{excluded}{the instrument types to exclude from the statistical outlier determination.}

\item{inferred}{the instrument types to flip the sign on in the the clean outlier function.}

\item{external}{T/F whether to apply the external override logic (see functions above).}

\item{tradeLvl}{T/F whether to explain the outliers via residual aggregate or trade level (see functions above).}

\item{params}{the remaining statistical outlier parameters (see functions above).}

\item{portfolios}{the rdecaf portfolios endpoint. Can be a subset.}

\item{session}{The rdecaf session}

\item{lookBack}{the backward time horizon from the date to be analyzed. Defaults to one year.}
}
\value{
A list with a detailed data frame and a summarized/aggregate data frame.
}
\description{
This is the description
}
