% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnl.R
\name{overwriteEvents}
\alias{overwriteEvents}
\title{Overwrites columns that have missing date from the pnl endpoint.}
\usage{
overwriteEvents(ledge, res, joinC = "id")
}
\arguments{
\item{ledge}{the list containing data frame being overwritten from flatten events.}

\item{res}{the resource data to feed quantity and asset class data columns.}

\item{joinC}{the column name used to join the data frames.}
}
\value{
A flat df containing the ledgers with the correct asset class info.
}
\description{
This is the description
}
