% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseDate}
\alias{parseDate}
\title{A function to transform arbitrary date formats into YYYY-MM-DD.}
\usage{
parseDate(str, origin = "1899-12-30")
}
\arguments{
\item{str}{A single date string.}

\item{origin}{The origin for numerical date inputs (default = "1899-12-30".}
}
\value{
A date value.
}
\description{
This is a description.
}
