% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseOffsetDateTime}
\alias{parseOffsetDateTime}
\title{A function to parse offset date time object.}
\usage{
parseOffsetDateTime(
  x,
  format = "\%Y-\%m-\%dT\%H:\%M:\%OS\%z",
  tz = "Asia/Singapore"
)
}
\arguments{
\item{x}{A vector with offset date.}

\item{format}{The format of the input.}

\item{tz}{The desired time zone.}
}
\value{
A list with the parsed date time.
}
\description{
This is a description.
}
