% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remapDB.R
\name{patchDataframe}
\alias{patchDataframe}
\title{This function puts and patches from source to target data-frame.}
\usage{
patchDataframe(targetDF, sourceDF, mField, putMap, patchMap)
}
\arguments{
\item{targetDF}{The target data frame.}

\item{sourceDF}{The source data frame.}

\item{mField}{The respective fields to be matched to find the identity of rows.}

\item{putMap}{The fields to be putted.}

\item{patchMap}{The fields to be patched.}
}
\value{
Returns a putted/patched target data frame.
}
\description{
This is a description.
}
