% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeorder.R
\name{patchTrdOrdExecutions}
\alias{patchTrdOrdExecutions}
\title{A function to patch trade order executions with the trade id.}
\usage{
patchTrdOrdExecutions(trdOrdExecutions, session)
}
\arguments{
\item{trdOrdExecutions}{A data-frame with the 'booking' column populated.}

\item{session}{The rdecaf session.}
}
\description{
This is the description
}
