% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{percentify}
\alias{percentify}
\title{A function for formatting percentages:}
\usage{
percentify(number, digit = 2)
}
\arguments{
\item{number}{An array of numeric values.}

\item{digit}{the number of digits. Default is 2.}
}
\value{
A string of percentage:
}
\description{
This is a description.
}
