% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance.AlignData}
\alias{performance.AlignData}
\title{A helper function to align 2 results the performance endpoint.}
\usage{
performance.AlignData(primary, secondary)
}
\arguments{
\item{primary}{The primary performance result.}

\item{secondary}{The secondary performance result.}
}
\value{
The secondary performance result aligne to the primary.
}
\description{
This is the description
}
