% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance.ComputeRelative}
\alias{performance.ComputeRelative}
\title{A function to compute the relative performance between 2 getPerformanceV2 results.}
\usage{
performance.ComputeRelative(primary, secondary)
}
\arguments{
\item{primary}{The primary getPerformanceV2 result.}

\item{secondary}{The secondary getPerformanceV2 result.}
}
\value{
The relative performance result.
}
\description{
This is the description
}
