% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performancefee.R
\name{performanceFeeAccountant}
\alias{performanceFeeAccountant}
\title{This function prepares the payload and syncs the PJE's to the performance fee account.}
\usage{
performanceFeeAccountant(x, session, aTypeName = "Performance Fees")
}
\arguments{
\item{x}{The data frame with following columns: "perfFeeTotal", "resmain", "pacc", "sc".}

\item{session}{The rdecaf session.}

\item{aTypeName}{The name of the analytical type.}
}
\description{
This is the description
}
