% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{performanceOutliers}
\alias{performanceOutliers}
\title{A function prepares the performance outliers data-frame}
\usage{
performanceOutliers(portfolio, start = NULL, factor = 8, session)
}
\arguments{
\item{portfolio}{The portfolio id,}

\item{start}{The start date of the performance. Default is start of the current year.}

\item{factor}{The factor to sd()}

\item{session}{The rdecaf session.}
}
\value{
A data-frame with outliers.
}
\description{
This is the description
}
