% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{periodDates}
\alias{periodDates}
\title{A function to produce dates in requested periodicity.}
\usage{
periodDates(
  periodLetter,
  yearsOfHistory = 5,
  endOfWeek = "Friday",
  date = Sys.Date()
)
}
\arguments{
\item{periodLetter}{Any of "D", "W", "M", "Q", "Y"}

\item{yearsOfHistory}{The amount of years to lookback. Default is 5.}

\item{endOfWeek}{For weekly periodicity, the end of week day. Default is "Friday".}

\item{date}{The date to traverse back from.}
}
\value{
A sequence of dates for periodicity.
}
\description{
This is a description.
}
