% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictetCommons.R
\name{pictetEBankingNormalizer}
\alias{pictetEBankingNormalizer}
\title{This function normalizes the position file from Pictet's ebanking system.}
\usage{
pictetEBankingNormalizer(filePath, session)
}
\arguments{
\item{filePath}{The data frame with flattened records.}

\item{session}{The rdecaf session.}
}
\value{
Returns the normalised records.
}
\description{
This is the description
}
