% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictetCommons.R
\name{pictetEBankTXNQuantizer}
\alias{pictetEBankTXNQuantizer}
\title{This function quantizes the transfer actions based on pictet's eBanking transactin file.}
\usage{
pictetEBankTXNQuantizer(file, resources, session)
}
\arguments{
\item{file}{The file path.}

\item{resources}{The resources data frame.}

\item{session}{The rdecaf session.}
}
\value{
Quantizes transfer actions. Return NULL
}
\description{
This is the description
}
