% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictetCommons.R
\name{pictetFileMappers}
\alias{pictetFileMappers}
\title{A function to return the ltype file mappers.}
\usage{
pictetFileMappers()
}
\value{
Returns a list with the file mappers.
}
\description{
This is the description
}
