% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictetCommons.R
\name{pictetFileNormalizer}
\alias{pictetFileNormalizer}
\title{A function to normalise the pictet flattened files for give ltype.}
\usage{
pictetFileNormalizer(data, ltype, accounts)
}
\arguments{
\item{data}{The data frame with flattened records.}

\item{ltype}{The ltype of the records.}

\item{accounts}{the account mapping.}
}
\value{
Returns the normalised records.
}
\description{
This is the description
}
