% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictetCommons.R
\name{pictetSecTypeMap}
\alias{pictetSecTypeMap}
\title{A function to return the pictet security type vs resource ctype mapping}
\usage{
pictetSecTypeMap()
}
\value{
Returns a vector with the mapping.
}
\description{
This is the description
}
