% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictetCommons.R
\name{pictetTXNCode}
\alias{pictetTXNCode}
\title{A function to return the transaction code vs trade ctype mapping.}
\usage{
pictetTXNCode()
}
\value{
Returns a vector with the mapping.
}
\description{
This is the description
}
