% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnl.R
\name{pnlContribReport}
\alias{pnlContribReport}
\title{Compiler fn that returns the data for the pnl contribution excel report.}
\usage{
pnlContribReport(portfolio, date, session)
}
\arguments{
\item{portfolio}{the decaf portfolio id.}

\item{date}{the end date for the pnl relative period calculation via compilepnlreport.}

\item{session}{the decaf session.}

\item{symbol}{string of the instrument symbol to debug. Defaults to NULL.}
}
\value{
A data frame of the relative period pnls in a consolidated contribution report.
}
\description{
This is the description
}
