% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnl.R
\name{pnlPreembleWrapper}
\alias{pnlPreembleWrapper}
\title{A wrapper function to get the pnl preemble.}
\usage{
pnlPreembleWrapper(portfolio, startDate, endDate, session)
}
\arguments{
\item{portfolio}{The portfolio id.}

\item{startDate}{The start date.}

\item{endDate}{The end date.}

\item{session}{The rdecaf session.}
}
\value{
A list with the PnL preemble
}
\description{
This is the description
}
