% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnl.R
\name{pnlReport}
\alias{pnlReport}
\title{Compiler fn that returns all elements to construct basic pnl report.}
\usage{
pnlReport(
  portfolio,
  since,
  until,
  currency,
  session,
  res,
  cashS = "CCY",
  transfS = "transfer|investment",
  depoS = "DEPO",
  apiURL = "/apis/function/valuation-reports/eventsreport"
)
}
\arguments{
\item{portfolio}{the decaf portfolio id.}

\item{since}{the start date for the pnl calculation.}

\item{until}{the end date for the pnl calculation.}

\item{currency}{the desired currency for the computations.}

\item{session}{the decaf session.}

\item{res}{the df containing the asset class elements to overwrite faulty data from getStocksAndAssets.}

\item{cashS}{the string to identify cash instrument types.}

\item{transfS}{the string(s) to identify transfer trade types.}

\item{depoS}{the string to identify deposit instrument types.}

\item{apiURL}{the npl endpoint API URL string.}
}
\value{
A list containing the the granular ledger info  and a summarized one liner position data frame of pnl agg info corrected.
}
\description{
This is the description
}
