% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{positionDifferential}
\alias{positionDifferential}
\title{This function prepares the position quantities of external vis-a-vis decaf stocks:}
\usage{
positionDifferential(data, session)
}
\arguments{
\item{data}{The data with the external positions. Expected columns: resmain, qtymain, pxmain, accmain}

\item{session}{the rdecaf session}
}
\value{
Returns a data frame with the respective quantities:
}
\description{
This is the description
}
