% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{positionPnLWrapper}
\alias{positionPnLWrapper}
\title{A wrapper function for calculating the position wise pnls.}
\usage{
positionPnLWrapper(holdings, holdingsParams, resources, ccy, date, session,
  portName, portfolio, begPeriod, fxobs)
}
\arguments{
\item{holdings}{The holdings list from getPrintableHoldings.}

\item{holdingsParams}{The list of paramters used for holdings.}

\item{resources}{The resources data frame.}

\item{ccy}{The currency of the container.}

\item{date}{The report date.}

\item{session}{The rdecaf session.}

\item{portName}{The container name.}

\item{portfolio}{The portfolio id.}

\item{begPeriod}{The beginning period.}

\item{fxobs}{The FX observatios for the same period.}
}
\value{
The period to date pnls.
}
\description{
This is the description
}
