% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocketBase.R
\name{postIt}
\alias{postIt}
\title{Pocketbase dependent function to push data in the DB. Ancillary to the incrIt fn below.}
\usage{
postIt(rb, path, lst)
}
\arguments{
\item{rb}{The rocketbase authorizaion credentials. Defined externally.}

\item{path}{The path of the pocketbase API endpoint, e.g. /api/collections/cfraEquities/records}

\item{lst}{list of the data to be pused to the endpoint above.}
}
\value{
NULL.
}
\description{
This is the description
}
