% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{postResourcesByBatch}
\alias{postResourcesByBatch}
\title{A function to post resources in batches.}
\usage{
postResourcesByBatch(resources, batchSize = 1000, session)
}
\arguments{
\item{resources}{The resource data frame.}

\item{batchSize}{The size of the batches.}

\item{session}{The DECAF session info.}
}
\value{
Returns NULL
}
\description{
This is the description
}
