% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav.R
\name{prefaceSeries}
\alias{prefaceSeries}
\title{Given a NAV or return time series df, flags extreme outliers before other treatment, given scale}
\usage{
prefaceSeries(series, scale = 5)
}
\arguments{
\item{series}{the time series data frame}

\item{scale}{the scale applied to IQR calculation to define outlier}
}
\value{
A data frame with the extreme outliers flagged.
}
\description{
This is the description
}
