% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{prepareAccountPayload}
\alias{prepareAccountPayload}
\title{This function prepares the account payload as data frame.}
\usage{
prepareAccountPayload(
  accounts,
  portfolios,
  institutions,
  rccy = "USD",
  atype = NA,
  session
)
}
\arguments{
\item{accounts}{The accounts list.}

\item{portfolios}{The corresponding portfolio id's in the oder of account list.}

\item{institutions}{The institutions data frame.}

\item{rccy}{The fallback value for the account currency.}

\item{atype}{The analytical type of account. Default is NA.}

\item{session}{The session.}
}
\value{
The accounts payload as data frame.
}
\description{
This is the description
}
