% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbg.R
\name{prepareBBGDLData}
\alias{prepareBBGDLData}
\title{Function to prepare ticker for request.}
\usage{
prepareBBGDLData(session, zero = Sys.Date(), date, underlying,
  fldsByCtype, other = FALSE, field = "symbol", ohlc = TRUE)
}
\arguments{
\item{session}{The decaf session.}

\item{zero}{Should closed stocks be included? 1 = yes, - = no.}

\item{date}{The date for the stocks.}

\item{underlying}{Should underlying instruments be included?}

\item{fldsByCtype}{The list with fields to be considered for each ctype.}

\item{other}{Shall OTHER assets be considered as well?}

\item{field}{The field which should be considered in decaf resources.}

\item{ohlc}{Should the ohlccode be considered as an id.}
}
\value{
A list with a) tickers for price, b) tickers for reference, c) the resource data frame
}
\description{
Function to prepare ticker for request.
}
