% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdings.R
\name{prepareHoldingsTables}
\alias{prepareHoldingsTables}
\title{This function prepares the holding details and summary tables for html/pdf.
It requires the holdings list from getPrintableHoldings.}
\usage{
prepareHoldingsTables(
  holdings,
  colSelect,
  colOverride = NULL,
  summaryCol = "Subtype",
  summaryKey = "HEBELE1234"
)
}
\arguments{
\item{holdings}{The holdings list from getPrintableHoldings.}

\item{colSelect}{Which columns from the getPrintableHoldings should be selected?}

\item{colOverride}{Custom column names. Default is NULL.}

\item{summaryCol}{The column of holdings by which to summarize the holdings by.}

\item{summaryKey}{Optionally, you may select sublevels for the summary for a key in summaryCol.}
}
\value{
A list with the holdingsDetails, holdingsSummary data frames, and the original holdings list.
}
\description{
This is the description
}
