% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figi.R
\name{prepareOpenFigiData}
\alias{prepareOpenFigiData}
\title{A function to prepare the data-frame for the figiCall function.}
\usage{
prepareOpenFigiData(df, idType, fld, ccy)
}
\arguments{
\item{df}{The data-frame}

\item{idType}{The type of ID to use for openFigi. Default is ID_ISIN.}

\item{fld}{The column name of the data-frame which has the corresponding instrument id.}

\item{ccy}{The column name for the currency of the instrument.}
}
\value{
Returns the openFigi comform data-frame.
}
\description{
This is a description.
}
