% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{productCodeFromTicker}
\alias{productCodeFromTicker}
\title{A function to infer the product symbol from ticker:}
\usage{
productCodeFromTicker(ticker)
}
\arguments{
\item{ticker}{The bloomberg ticker.}
}
\value{
Returns the inferred product code of a bloomber ticker.
}
\description{
This is a description.
}
