% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{pushOhlc}
\alias{pushOhlc}
\title{A function to push ohlc observations}
\usage{
pushOhlc(symbol, close, date, session, source = NA)
}
\arguments{
\item{symbol}{The vector of symbols.}

\item{close}{The vector of close values.}

\item{date}{The vector of dates.}

\item{session}{The rdecaf session.}

\item{source}{The source. Default NA.}
}
\value{
Returns NULL
}
\description{
This is the description
}
