% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionIncrement.R
\name{pushPosDifferences}
\alias{pushPosDifferences}
\title{This function pushes the positions difference between decaf and provider}
\usage{
pushPosDifferences(stocks, provider, session, ctype = 20, invert = TRUE)
}
\arguments{
\item{stocks}{The position comparisons data frame.}

\item{provider}{The provider as string.}

\item{session}{The rdecaf session.}

\item{ctype}{The ctype. Default is 20.}

\item{invert}{Shall qty/px be inverted if px is negative? Default =TRUE}
}
\description{
This is the description
}
