% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{quantityFinder}
\alias{quantityFinder}
\title{A function to infer the rounded factor between two values.}
\usage{
quantityFinder(sourceValue, targetValue)
}
\arguments{
\item{sourceValue}{The value to compute the factor for.}

\item{targetValue}{The value to compute the factor from.}
}
\value{
Returns a factor.
}
\description{
This is a description.
}
