% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfra.R
\name{queryCfra}
\alias{queryCfra}
\title{CFRA-pocketbase dependent function to query the CFRA endpoint, given client authorization config.}
\usage{
queryCfra(client, endpoint)
}
\arguments{
\item{client}{The authorization config as defined outside. Should be an R6Class object.}

\item{endpoint}{The respective CFRA endpoint, e.g. 'equities'. See documentation here https://developers.cfraresearch.com/.}
}
\value{
list of the endpoint contents.
}
\description{
This is the description
}
