% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{queueIncoming}
\alias{queueIncoming}
\title{A function to queue incoming files to be considered for file pipeline.}
\usage{
queueIncoming(mapper)
}
\arguments{
\item{mapper}{The income file mapper.}
}
\value{
Produces a log of the queueing process and writes to the log of particular session.
}
\description{
This is the description
}
