% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{recodeIt}
\alias{recodeIt}
\title{A function to overwrite variables in dataframe according to an outline.}
\usage{
recodeIt(data, bp, var, sweep = FALSE)
}
\arguments{
\item{data}{data frame containing the variable to overwrite, e.g. resources.}

\item{bp}{list dictating the blueprint of the overwriting function that will be converted into expression. See below.}

\item{var}{vector of the variable name to be changed, e.g. assetclass.}

\item{sweep}{logical to dictate whether to overwrite values with existing values or not. Defaults to false.}
}
\value{
data frame with the revised column - as character - and an added flag column 'recode' indicating where changes were made.
}
\description{
This is the description
}
\examples{
bp <- list(
"5" ="ctype=='CCY'",
"6" ="ctype \%in\% c('LOAN', 'DEPO', 'COD')",
"13"="ctype=='BOND'",
"17"="ctype \%in\% c('FXFWD', 'FXOPT','DCIP')",
"27"="ctype=='INSR'",
"11"="ctype=='ELNP'",
"9" ="ctype=='SHRE' & safeGrep(trimConcatenate(`attributes.figistype`), 'COMMONSTOCK') == '1'"
) 
}
