% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbg.R
\name{referenceFieldRequestBuilder}
\alias{referenceFieldRequestBuilder}
\title{A function to build the reference data request object.}
\usage{
referenceFieldRequestBuilder(resources, bbgFields, forceAux = FALSE)
}
\arguments{
\item{resources}{The resources data frame.}

\item{bbgFields}{The list of bbg field mappings.}

\item{forceAux}{Are the aux fields to be forced?}
}
\value{
Either a list with the reference data object or NULL.
}
\description{
A function to build the reference data request object.
}
