% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{regressions}
\alias{regressions}
\title{Returns list of regression df and associated ts df for one factor lms only}
\usage{
regressions(
  y,
  x,
  tsColname = "date",
  rgColname = "close",
  minobs = 30,
  roll = 5,
  stdev = 3
)
}
\arguments{
\item{y}{the df containing the lhs time series to be regressed}

\item{x}{the df containing the rhs times series}

\item{tsColname}{the column name of the time series date value, defaults to date}

\item{rgColname}{the column name of the time series value, defaults to close}

\item{minobs}{the min #obs to run a regression, defaults to 30}

\item{roll}{the rolling avg lag to apply, defaults to 5 days}

\item{stdev}{the # of st.devs to use when trimming outliers, defaults to 3}
}
\value{
A list with model alpha/beta info and associated filtered time series dfs.
}
\description{
This is a description.
}
