% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{renderTemplate}
\alias{renderTemplate}
\title{Renders an html template.}
\usage{
renderTemplate(tmplPath, ...)
}
\arguments{
\item{tmplPath}{The path to the template.}

\item{...}{Additional parameters to be passed to tmpl.}
}
\value{
An html.
}
\description{
This is a description.
}
