% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbg.R
\name{requestableTickers}
\alias{requestableTickers}
\title{Function to prepare ticker for request.}
\usage{
requestableTickers(session, date, zero, underlying, fldsByCtype,
  other = FALSE)
}
\arguments{
\item{session}{The decaf session.}

\item{date}{The date to query stocks.}

\item{zero}{Should closed stocks be included? 1 = yes, - = no.}

\item{underlying}{Should underlying instruments be included?}

\item{fldsByCtype}{The list with fields to be considered for each ctype.}

\item{other}{Shall OTHER assets be considered as well?}
}
\value{
A list with a) tickers for price, b) tickers for reference, c) the resource data frame
}
\description{
Function to prepare ticker for request.
}
