% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionIncrement.R
\name{resolvePositions}
\alias{resolvePositions}
\title{This function computes the position differences and optionally reconciles in the system.}
\usage{
resolvePositions(
  accWisePos,
  instName,
  resources,
  date,
  ctypes,
  reconcile,
  tradeType,
  session
)
}
\arguments{
\item{accWisePos}{A list with account wise positions.}

\item{instName}{The name of the institution to be used for recons trade guid.}

\item{resources}{The resources data frame.}

\item{date}{The date of comparison.}

\item{ctypes}{The resource ctypes to be compared.}

\item{reconcile}{Should the differences be reconciled.}

\item{tradeType}{The ctype of the recons trades.}

\item{session}{The rdecaf session.}
}
\description{
This is the description
}
