% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{retreatReport}
\alias{retreatReport}
\title{Defines a function to retreat the report data:}
\usage{
retreatReport(indata)
}
\arguments{
\item{indata}{Data as returned by `getPrintableHoldings` function.}
}
\value{
Retreated report data:
}
\description{
Defines a function to retreat the report data:
}
