% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{returnOutliers}
\alias{returnOutliers}
\title{A function to detect outliers in xts return series:}
\usage{
returnOutliers(xtsReturns, factor)
}
\arguments{
\item{xtsReturns}{Return series of xts.}

\item{factor}{The factor of stdev to apply}
}
\value{
A vector with TRUE/FALSE values
}
\description{
This is the description
}
