% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{rSync}
\alias{rSync}
\title{A function to call the system command rsync}
\usage{
rSync(server, location, target, folderNames, exclude = "")
}
\arguments{
\item{server}{The server address}

\item{location}{The location on the server.}

\item{target}{The target directory locally.}

\item{folderNames}{The names of the local folders.}

\item{exclude}{A vector with keys to exclude from syncin.}
}
\value{
A message to show the log of the rsync call.
}
\description{
This is the description
}
