% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clever.R
\name{runCleverFunction}
\alias{runCleverFunction}
\title{The clever wrapper for running a function.}
\usage{
runCleverFunction(func, client, session, config, job, arguments)
}
\arguments{
\item{func}{The function to run.}

\item{client}{The list with the client info.}

\item{session}{The rdecaf session.}

\item{config}{The list with the clever configuration.}

\item{job}{The job information}

\item{arguments}{The list with workhorse arguments.}
}
\value{
NULL. Email with the alert will be sent.
}
\description{
This is the description
}
