% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{runDemail}
\alias{runDemail}
\title{A function runs the java demail.}
\usage{
runDemail(address, jarPath, host, port, folder, directory, archive)
}
\arguments{
\item{address}{The address of the email server.}

\item{jarPath}{The path of the demail jar file.}

\item{host}{The name of the email host.}

\item{port}{The port of the email server.}

\item{folder}{The name of the folder on the email server.}

\item{directory}{The local directory to sync to.}

\item{archive}{The name of the archive folder on server.}
}
\value{
A message to show the log of the demail process.
}
\description{
This is the description
}
