% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safeColumn}
\alias{safeColumn}
\title{A function to safely call a column in a data-frame whether column
exists or not.}
\usage{
safeColumn(df, col)
}
\arguments{
\item{df}{The data-frame.}

\item{col}{The expected column name.}
}
\value{
The column values or NA.
}
\description{
This is a description.
}
