% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safeCondition}
\alias{safeCondition}
\title{A function to whether values in data frame column
satisfies a condition.}
\usage{
safeCondition(df, col, condition)
}
\arguments{
\item{df}{The data frame.}

\item{col}{The column to be checked.}

\item{condition}{The condition to be checked.}
}
\value{
A vector with TRUE or FALSE
}
\description{
A function to whether values in data frame column
satisfies a condition.
}
