% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safeGrep}
\alias{safeGrep}
\title{A function to count keys in strings.}
\usage{
safeGrep(str, key)
}
\arguments{
\item{str}{A vector with strings.}

\item{key}{Key to be found.}
}
\value{
A vector of counts.
}
\description{
This is a description.
}
