% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safeMatch}
\alias{safeMatch}
\title{A function to safely check the condition of a vector of a string.}
\usage{
safeMatch(str, key)
}
\arguments{
\item{str}{A vector with strings.}

\item{key}{Key to be found.}
}
\value{
A vector with "1" for TRUE and "0" for FALSE.
}
\description{
This is a description.
}
