% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safeNull}
\alias{safeNull}
\title{A function to replace null and length 0 value with NA}
\usage{
safeNull(x)
}
\arguments{
\item{x}{A variable}
}
\value{
Returns x or NA.
}
\description{
This is the description
}
