% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safeRbind}
\alias{safeRbind}
\title{A function to safely rbind a list with data frames.}
\usage{
safeRbind(x, convertToString = FALSE)
}
\arguments{
\item{x}{A list with data frames.}

\item{convertToString}{Shall all values be converted to strings. Default is FALSE.}
}
\value{
A single data frame.
}
\description{
This is a description.
}
