% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safeTry}
\alias{safeTry}
\title{If try-error, return NA:}
\usage{
safeTry(x, nullToNa = TRUE)
}
\arguments{
\item{x}{A object of try.}

\item{nullToNa}{Boolean to indicate whether null should be replaced to NA. Default is TRUE.}
}
\value{
NA or valid value.
}
\description{
If try-error, return NA:
}
