% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarasinCommons.R
\name{sarasinCTypeMapper}
\alias{sarasinCTypeMapper}
\title{A function to return the sarasin security type vs resource ctype mapping}
\usage{
sarasinCTypeMapper()
}
\value{
Returns a vector with the mapping.
}
\description{
This is the description
}
