% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarasinCommons.R
\name{sarasinFlattenM001}
\alias{sarasinFlattenM001}
\title{A function to flatten M001 files.}
\usage{
sarasinFlattenM001(data)
}
\arguments{
\item{data}{The sarasin M005 data}
}
\value{
Returns a flat file
}
\description{
This is the description
}
