% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarasinCommons.R
\name{sarasinFlattenM005}
\alias{sarasinFlattenM005}
\title{A function to flatten M005 files.}
\usage{
sarasinFlattenM005(data)
}
\arguments{
\item{data}{The sarasin M005 data}
}
\value{
Returns a flat file
}
\description{
This is the description
}
