% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarasinCommons.R
\name{sarasinSTypeMapper}
\alias{sarasinSTypeMapper}
\title{A function to return the sarasin security type vs resource stype mapping}
\usage{
sarasinSTypeMapper()
}
\value{
Returns a vector with the mapping.
}
\description{
This is the description
}
