% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figi.R
\name{securityTypeCurrency}
\alias{securityTypeCurrency}
\title{A function to provide the figi type currency mapping:}
\usage{
securityTypeCurrency()
}
\value{
Returns a matrix with the figi type currency mapping.
}
\description{
This is a description.
}
