% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnl.R
\name{selPnlContrCols}
\alias{selPnlContrCols}
\title{Helper fn to clean up the PnL output.}
\usage{
selPnlContrCols(
  df,
  key = "symbol",
  sfx = "(itd)",
  cols = c("unrealised", "realised", "income", "fees", "total", "total gross",
    "total net", "contr"),
  exc = NULL
)
}
\arguments{
\item{df}{the pnl df, usually from the relative series, e.g. ytd.}

\item{key}{string of the column used for joining all the dfs. Defaults to symbol.}

\item{sfx}{string of relative series suffix to add to make the column names unique. Defaults to (itd) for inception to date.}

\item{cols}{string vector of the column names to add a suffix to. See below.}

\item{exc}{string of the column name to exclude. Defaults to no exclusions.}
}
\value{
A cleaned-up data frame of pnl relative series.
}
\description{
This is the description
}
