% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{separateDFByTeam}
\alias{separateDFByTeam}
\title{A function to extract a data frame to a list based on ownership.}
\usage{
separateDFByTeam(df, fld, session)
}
\arguments{
\item{df}{The data frame.}

\item{fld}{The column name where account id information is.}

\item{session}{The rdecaf session.}
}
\value{
A team wise list with the segregated data frame, team name and users.
}
\description{
This is the description
}
