% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{sepCfrmP}
\alias{sepCfrmP}
\title{A function to derive calls vs puts for option instruments.}
\usage{
sepCfrmP(
  res,
  ctypeF = "OPT",
  cp = c("C", "P"),
  xtractFirst = c("EQUITY"),
  xtractLast = c("COMDTY", "INDEX"),
  nameMatch = "\\\\b(CALL|PUT|CALLS|PUTS)\\\\b",
  sweep = FALSE
)
}
\arguments{
\item{res}{data frame containing the resources endpoint}

\item{ctypeF}{string to filter out for ctype. Defaults to OPT.}

\item{cp}{string vector dictating how calls and puts are identified. Defaults to C and P.}

\item{xtractFirst}{string vector identifying the instrument symbol will contain this string, where the above letters are extracted from the first char. Defaults to Equity.}

\item{xtractLast}{string vector identifying the instrument symbol will contain this string, where the above letters are extracted from the last char. Defaults to commodity and index.}

\item{nameMatch}{regex string vector identifying the patterns to look for in the instrument name for symbols beyond the above. Defaults to call(s) or put(s).}

\item{sweep}{logical to dictate whether to overwrite values with existing values or not. Defaults to false.}
}
\value{
data frame with the cleaned up underlying column.
}
\description{
This is the description
}
