% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setColsToNull}
\alias{setColsToNull}
\title{A function to set specific columns to NULL}
\usage{
setColsToNull(df, cols)
}
\arguments{
\item{df}{A data-frame.}

\item{cols}{A vector with the selective column names.}
}
\value{
A data-frame.
}
\description{
This is a description.
}
