% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setNAColTo}
\alias{setNAColTo}
\title{Set NA's for columns in data frame to a desired value.}
\usage{
setNAColTo(df, cols, val = 0)
}
\arguments{
\item{df}{The data frame.}

\item{cols}{A vectors of column names.}

\item{val}{The value NA's should be changed to. Default is 0.}
}
\value{
A data frame.
}
\description{
This is the description
}
