% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{settleFuturePnL}
\alias{settleFuturePnL}
\title{This function creates a ficticious PnL account and settles the Future pnl T-1}
\usage{
settleFuturePnL(futurePositions, accounts, resources, session)
}
\arguments{
\item{futurePositions}{The openPositions data frame coming from getOpenpositionbyctype.}

\item{accounts}{The accounts data frame from rdecaf.}

\item{resources}{The resources data frame.}

\item{session}{The rdecaf session.}
}
\description{
This is the description
}
