% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sourceFromFolder}
\alias{sourceFromFolder}
\title{A Wrapper function to source R scripts from specified paths.}
\usage{
sourceFromFolder(paths, pattern = "*.R")
}
\arguments{
\item{paths}{The path where the R scripts are located.}

\item{pattern}{The pattern of the R scripts (default = *.R).}
}
\value{
Source R scripts:
}
\description{
A Wrapper function to source R scripts from specified paths.
}
