% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{specialCharacterTreater}
\alias{specialCharacterTreater}
\title{A function to treat special characters:}
\usage{
specialCharacterTreater(df, cols = NULL)
}
\arguments{
\item{df}{A data-frame.}

\item{cols}{A vector with the selective column names to be treated.}
}
\value{
A data-frame.
}
\description{
This is a description.
}
