% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocketBase.R
\name{stackData}
\alias{stackData}
\title{Pocketbase dependent function to query a rocketbase endpoint and stack the results X pages at a time.}
\usage{
stackData(pages, rb, path)
}
\arguments{
\item{pages}{The number of pages corresponding to the endpoint. We can only query one page at a time.}

\item{rb}{The rocketbase authorizaion credentials. Defined externally.}

\item{path}{The path of the pocketbase API endpoint, e.g. /api/collections/cfraEquities/records}
}
\value{
a data frame of the stacked records.
}
\description{
This is the description
}
