% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{stackedBarChart}
\alias{stackedBarChart}
\title{Provides a stacked bar chart.}
\usage{
stackedBarChart(df, colors = RColorBrewer::brewer.pal(9, "GnBu")[3:9])
}
\arguments{
\item{df}{A data frame. df[, 1] == label, df[,2] == value}

\item{colors}{The color palette to be used.}
}
\value{
A stacked bar chart.
}
\description{
This is the description
}
