% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{stackedBarChartTranspose}
\alias{stackedBarChartTranspose}
\title{Provides a stacked bar chart.}
\usage{
stackedBarChartTranspose(df, barColor = NA, labelColor = NA)
}
\arguments{
\item{df}{A data frame. df[, 1] == label, df[,2] == value}

\item{barColor}{The color of the bars.}

\item{labelColor}{The color of the lables.}
}
\value{
A stacked bar chart.
}
\description{
This is the description
}
