% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stringCleaner}
\alias{stringCleaner}
\title{A function to gsub strings.}
\usage{
stringCleaner(string, cleaningMaps = c(`&#62;LINK&#60;/a&#62;` =
  ">LINK</a>", `&#60;a href` = "<a href"))
}
\arguments{
\item{string}{The original string.}

\item{cleaningMaps}{The vector with the pattern and the replacement.}
}
\value{
The cleaned string,
}
\description{
A function to gsub strings.
}
