% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stringMatch}
\alias{stringMatch}
\title{A function to match a string with candidates:}
\usage{
stringMatch(sourceStr, targetStr, jaccard = 0.15)
}
\arguments{
\item{sourceStr}{A vector with strings to match from.}

\item{targetStr}{A vector with strings to match against.}

\item{jaccard}{The acceptable threshold for the jaccard coefficient.}
}
\value{
A vector with matching target indices:
}
\description{
A function to match a string with candidates:
}
