% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{sumStats}
\alias{sumStats}
\title{Simplifies the compute returns stats advanced stats into the key metrics for the period in simple df format}
\usage{
sumStats(df, suffix = "")
}
\arguments{
\item{df}{the complex computereturnstats data frame input.}

\item{suffix}{the suffix string to provide the row names, e.g. benchmark.}
}
\value{
A simplified data frame containing return, volatility and sharpe metrics.
}
\description{
This is the description
}
