% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{syncAccounts}
\alias{syncAccounts}
\title{A function to sync accounts between 2 DECAF instances:}
\usage{
syncAccounts(sourceSession, targetSession, atypes, institutions,
  portfolios, accounts)
}
\arguments{
\item{sourceSession}{The session info for the DECAF instance to sync from.}

\item{targetSession}{The session info for the DECAF instance to sync to.}

\item{atypes}{The data-frame with analtyical types.}

\item{institutions}{The data-frame with institutions.}

\item{portfolios}{The data-frame with portfolios.}

\item{accounts}{The data-frame with the source accounts.}
}
\value{
The function creates the accounts in the target instance and returns a accounts data-frame.
}
\description{
This is the description
}
