% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{syncAgent}
\alias{syncAgent}
\title{A function to sync agents between 2 DECAF instances:}
\usage{
syncAgent(sourceSession, targetSession, tradeAgents)
}
\arguments{
\item{sourceSession}{The session info for the DECAF instance to sync from.}

\item{targetSession}{The session info for the DECAF instance to sync to.}

\item{tradeAgents}{The agent id's coming from the source trades.}
}
\value{
The function creates the agents in the target instance and returns a agent class data-frame.
}
\description{
This is the description
}
