% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{syncAtypes}
\alias{syncAtypes}
\title{A function to sync analytical types between 2 DECAF instances:}
\usage{
syncAtypes(sourceSession, targetSession, sourceAccounts)
}
\arguments{
\item{sourceSession}{The session info for the DECAF instance to sync from.}

\item{targetSession}{The session info for the DECAF instance to sync to.}

\item{sourceAccounts}{The accounts data-frame coming from the source instance.}
}
\value{
The function creates the analytical types in the target instance and returns a analytical types data-frame.
}
\description{
This is the description
}
