% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{syncExtValuation}
\alias{syncExtValuation}
\title{A function to sync external valuations between 2 DECAF instances:}
\usage{
syncExtValuation(
  sourceSession,
  targetSession,
  targetPortfolios,
  targetShareclasses,
  accounts
)
}
\arguments{
\item{sourceSession}{The session info for the DECAF instance to sync from.}

\item{targetSession}{The session info for the DECAF instance to sync to.}

\item{targetPortfolios}{The data-frame with target portfolios data frame.}

\item{targetShareclasses}{The data-frame with target share classes data frame.}

\item{accounts}{The data-frame with the source accounts.}
}
\value{
The function creates the accounts in the target instance and returns a accounts data-frame.
}
\description{
This is the description
}
