% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{syncOHLC}
\alias{syncOHLC}
\title{A function to sync ohlc observations between 2 DECAF instances.}
\usage{
syncOHLC(sourceSession, targetSession, resources, lte = NULL,
  lookBack = NULL, targetPrefix = NULL)
}
\arguments{
\item{sourceSession}{The session info for the DECAF instance to sync from.}

\item{targetSession}{The session info for the DECAF instance to sync to.}

\item{resources}{The resource data frame.}

\item{lte}{TODO:}

\item{lookBack}{TODO:}

\item{targetPrefix}{TODO:}
}
\value{
TODO:
}
\description{
This is the description
}
