% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{syncPortfolios}
\alias{syncPortfolios}
\title{A function to sync portfolios between 2 DECAF instances.}
\usage{
syncPortfolios(sourceSession, targetSession, accounts, teamGUID,
  exclude = NULL, search = NULL, customName = NULL,
  sourcePortfolios = NULL)
}
\arguments{
\item{sourceSession}{The session info for the DECAF instance to sync from.}

\item{targetSession}{The session info for the DECAF instance to sync to.}

\item{accounts}{The account/portfolio mapper between the 2 instances.}

\item{teamGUID}{The team GUID of the target instance.}

\item{exclude}{A vector with key words to exclude unwanted columns in source portfolios.}

\item{search}{A key word to search for. Default is NULL.}

\item{customName}{A vector of strings to replace the names of source portfolios with.}

\item{sourcePortfolios}{The rdecaf portfolio data frame from source.}
}
\value{
The function creates the portfolios in the target instance and return the source portfolio data-frame.
}
\description{
This is the description
}
