% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{syncShareclass}
\alias{syncShareclass}
\title{A function to sync share classes between 2 DECAF instances.}
\usage{
syncShareclass(sourceSession, targetSession, sourcePortfolios)
}
\arguments{
\item{sourceSession}{The session info for the DECAF instance to sync from.}

\item{targetSession}{The session info for the DECAF instance to sync to.}

\item{sourcePortfolios}{The data-frame with the source portfolios.}
}
\value{
The function creates the share classes in the target instance and returns a share class data-frame.
}
\description{
This is the description
}
