% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{syncTrades}
\alias{syncTrades}
\title{A function to sync trades between 2 DECAF instances:}
\usage{
syncTrades(targetSession, trades, atypes, shrclss, institutions, agents)
}
\arguments{
\item{targetSession}{The session info for the DECAF instance to sync to.}

\item{trades}{The trades from the target instance.}

\item{atypes}{The data-frame with analtyical types.}

\item{shrclss}{The data-frame with the share classes.}

\item{institutions}{The data-frame with institutions.}

\item{agents}{The data-frame with agents.}
}
\value{
The function creates the trades in the target instance and returns a trades data-frame.
}
\description{
This is the description
}
