% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{syncTradesByBatch}
\alias{syncTradesByBatch}
\title{A function to sync trades in batches}
\usage{
syncTradesByBatch(trades, session, batchSize = 500)
}
\arguments{
\item{trades}{The trades data frame.}

\item{session}{The DECAF session info}

\item{batchSize}{The desired batch sizes. Default=150.}
}
\description{
This is the description
}
