% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{syncUpdateEmail}
\alias{syncUpdateEmail}
\title{A function to send system emails using the an HTML email template.}
\usage{
syncUpdateEmail(template, updateText, emailParams, timezone = "CET")
}
\arguments{
\item{template}{The HTML template.}

\item{updateText}{A list with the keyword and the corresponding replacement string.}

\item{emailParams}{The parameter list for the email report function.}

\item{timezone}{The timezone to be used. Default is CET.}
}
\value{
A java-object of class org.apache.commons.mail.SimpleEmail
}
\description{
The function searces for keywords in an HTML template and replaces
the keywords with some other HTML string.
}
