% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{timeSeriesPlot}
\alias{timeSeriesPlot}
\title{Provides an line time series plot.}
\usage{
timeSeriesPlot(
  df,
  smooth = 0.3,
  limitFactor = 0.05,
  title,
  ylab = "",
  col = "darkslategray3"
)
}
\arguments{
\item{df}{A data frame with 'value' and 'date' columns.}

\item{smooth}{The smoothing parameter.}

\item{limitFactor}{The factor with which to expand the limits.}

\item{title}{The title of the plot.}

\item{ylab}{The y-axis name.}

\item{col}{The color for the polygon.}
}
\value{
An line time series plot.
}
\description{
This is the description
}
