% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health.R
\name{tradePxDubiosity}
\alias{tradePxDubiosity}
\title{This function is an iteration of the subset function above, specifc for trades endpoint.}
\usage{
tradePxDubiosity(
  session,
  filt = "!resmain_ctype \%in\% c('DEPO','LOAN','CCY') & round(as.numeric(pxmain))<=1",
  params = "list(commitment__gte=Sys.Date()-365)",
  lookBack = 365,
  tolerance = 0.5,
  addLink = TRUE
)
}
\arguments{
\item{session}{the rdecaf session.}

\item{filt}{string of the filter expression to evaluate after pulling the endpoint data. Defaults to non-cash, depo, loan ctypes and pxvalues <1.}

\item{params}{string of the list expressing to evaluate inside addParams call to getdboject. Defaults to last year of relative data.}

\item{lookBack}{numerical value of days to lookback when querying the ohlc data to calculate moving average used for deviation derivation. Defaults to 1 year.}

\item{tolerance}{numerical value indicating threshold above which to return records that deviate too far from a moving average. Defaults to .5.}

\item{addLink}{logical indicating whether to add the decaf link. Defaults to TRUE, requires that the endpoint data frame have an 'id' column.}
}
\value{
The data frame containing relevant stylized records.
}
\description{
This is a description
}
