% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{treatPriceSeries}
\alias{treatPriceSeries}
\title{This function removes NA's and optionally detects and remove outliers.}
\usage{
treatPriceSeries(df, dCol, pCol, quantile = 0.998, surpressPlot = FALSE)
}
\arguments{
\item{df}{The data frame with the price and date columns.}

\item{dCol}{The column name of the date.}

\item{pCol}{The column name of the price.}

\item{quantile}{The quantile for the outlier detection}

\item{surpressPlot}{TODO}
}
\value{
Cleaned price series
}
\description{
This function removes NA's and optionally detects and remove outliers.
}
