% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav.R
\name{treatSeries}
\alias{treatSeries}
\title{Compiler function that runs find and clean outlier functions above for a given data frame of returns}
\usage{
treatSeries(
  factor = c(8, 3.5),
  df,
  session,
  resources,
  correct = 0.5,
  thresh = 0.02,
  toler = 0.2,
  portfolio = NULL,
  date = NULL,
  xVal = FALSE,
  currency = NULL,
  excluded = c("FUT", "FXFWD"),
  inferred = c("LOAN"),
  tradeLvl = FALSE
)
}
\arguments{
\item{factor}{vector of numeric factors c(upper,lower) used in standard deviation calculation to limit outlier derivation iteration start to end}

\item{df}{the returns (NAV) data frame}

\item{session}{The rdecaf session}

\item{resources}{the rdecaf resources endpoint. Not required, defaults to NULL}

\item{correct}{magnitude factor applied in calculation used to derive relevancy of outlier residual.}

\item{thresh}{magnitude factor applied in calculation used to derive relevancy of outlier residual.}

\item{toler}{buffer factor applied in calculation used to derive relevancy of outlier residual.}

\item{portfolio}{The portfolio id. Not required, defaults to NULL, taken from outliers if not found}

\item{date}{The outlier date. Not required, defaults to NULL, taken from outliers if not found}

\item{xVal}{true/false to dictate whether external valuation overrides should ignore outlier derivation}

\item{currency}{The outlier currency. Not required, defaults to NULL, taken as reference currency if not found}

\item{excluded}{vector defining resources (from quants endpoint) to exclude in analysis, e.g. fx future. Defaults to NULL}

\item{inferred}{vector defining resources (from quants endpoint) to switch the sign of, e.g. loan. Defaults to NULL}

\item{tradeLvl}{true/false to dictate whether residual should be calculated in aggregate or trade level. Aggregate by default}
}
\value{
A data frame with the outlier dates and details.
}
\description{
This is the description
}
