% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trimConcatenate}
\alias{trimConcatenate}
\title{A function to trim string and transform to upper case.}
\usage{
trimConcatenate(str)
}
\arguments{
\item{str}{A string.}
}
\value{
Returns the transformed string.
}
\description{
This is a description.
}
