% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trimDws}
\alias{trimDws}
\title{A function to trim double white spaces.}
\usage{
trimDws(str)
}
\arguments{
\item{str}{A vector of strings.}
}
\value{
A vector of strings with double white space removed.
}
\description{
A function to trim double white spaces.
}
