% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trimExcessWs}
\alias{trimExcessWs}
\title{A function to trim excess white spaces.}
\usage{
trimExcessWs(str)
}
\arguments{
\item{str}{A string.}
}
\value{
The trimmed string.
}
\description{
This is a description.
}
