% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{updateCorporateActionsDoc}
\alias{updateCorporateActionsDoc}
\title{A function to create/update the corporate actions document for resources for a specific instance.}
\usage{
updateCorporateActionsDoc(
  session,
  ctypes,
  commitment__gte = dateOfPeriod("Y-3")
)
}
\arguments{
\item{session}{The session of the instance.}

\item{ctypes}{The trade ctypes. Either '81', i.e Share Dividend, or '101', Bond Coupon, or both. Default c(81, 101).}

\item{commitment__gte}{Since which date should the corporate actions be consdered? When NULL for all history. Default is dateOfPeriod(Y-0).}
}
\value{
NULL. The function creates/updates corporate actions documents for resources, whereby ctype is: system+decaf://corporate-actions.
}
\description{
This is the description
}
