% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{userMultipleCountryLogin}
\alias{userMultipleCountryLogin}
\title{A function to infer dubious user login behaviour.}
\usage{
userMultipleCountryLogin(data, configJSON)
}
\arguments{
\item{data}{The data frame.}

\item{configJSON}{The json config file}
}
\value{
A user wise list with the user login analysis.
}
\description{
A function to infer dubious user login behaviour.
}
