% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valueOfNearestDate}
\alias{valueOfNearestDate}
\title{A function to give the value of the nearest date.}
\usage{
valueOfNearestDate(targetDate, data, tolerance, dateField = "date",
  valueField = "price")
}
\arguments{
\item{targetDate}{The target date}

\item{data}{A data frame with the date and value column.}

\item{tolerance}{The maximum allowed distance in days.}

\item{dateField}{The column name of the date in data.}

\item{valueField}{The column name of the value field.}
}
\value{
A list with date and value.
}
\description{
This is a description.
}
