% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volksbankCommons.R
\name{volksbankPosColumnMappers}
\alias{volksbankPosColumnMappers}
\title{A function to return the column mappers}
\usage{
volksbankPosColumnMappers()
}
\value{
Returns a list with the column mappers.
}
\description{
This is the description
}
