% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volksbankCommons.R
\name{volksbankSecTypeMapper}
\alias{volksbankSecTypeMapper}
\title{A function to return the security type mapper}
\usage{
volksbankSecTypeMapper()
}
\value{
Returns a list with the security type mappers.
}
\description{
This is the description
}
