% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpbankCommons.R
\name{vpbankPOSColumnMapper}
\alias{vpbankPOSColumnMapper}
\title{A function to return the column mappers}
\usage{
vpbankPOSColumnMapper()
}
\value{
Returns a list with the column mappers.
}
\description{
This is the description
}
