% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpbankCommons.R
\name{vpbankSecTypeMapper}
\alias{vpbankSecTypeMapper}
\title{A function to return the security type mapper}
\usage{
vpbankSecTypeMapper(reference)
}
\arguments{
\item{reference}{VP Bank's instrument reference data frame.}
}
\value{
Returns a list with the security type mappers.
}
\description{
This is the description
}
