% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{weightedPriceSeries}
\alias{weightedPriceSeries}
\title{A function to compare NAV's between 2 decaf systems.}
\usage{
weightedPriceSeries(meta, session, push = FALSE)
}
\arguments{
\item{accounts}{The account mapping list.}

\item{type}{The type of container. Either 'accounts' or 'portfolios'.}

\item{ccy}{The currency to be valued.}
}
\value{
A data frame with the NAv comparisons.
}
\description{
A function to compare NAV's between 2 decaf systems.
}
