% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnl.R
\name{wrapEvents}
\alias{wrapEvents}
\title{Wrapper function to return consumable PnL data.}
\usage{
wrapEvents(pnlst, res)
}
\arguments{
\item{pnlst}{the pnl list of data elements from the return value of bare get.}

\item{res}{the df containing the asset class elements to overwrite faulty data from getStocksAndAssets.}
}
\value{
A list containing the the granular ledger info from the above and a summarized one liner position data frame of pnl agg info.
}
\description{
This is the description
}
