% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{writeCell}
\alias{writeCell}
\title{Writes a value to cell to the sheet[row, col] and styles it.}
\usage{
writeCell(sheet, row, col, value, style)
}
\arguments{
\item{sheet}{Sheet to write the cell to.}

\item{row}{Row address of the cell.}

\item{col}{Col address of the cell.}

\item{value}{Cell value.}

\item{style}{Cell style.}
}
\value{
written cell.
}
\description{
Writes a value to cell to the sheet[row, col] and styles it.
}
