% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{writeFundReport}
\alias{writeFundReport}
\title{Writes the fund report.}
\usage{
writeFundReport(report, file)
}
\arguments{
\item{report}{Report data.}

\item{file}{Path to file to write the excel to.}
}
\value{
The workbook written to the file.
}
\description{
Writes the fund report.
}
