% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{writeTableHeader}
\alias{writeTableHeader}
\title{Writes table header.}
\usage{
writeTableHeader(sheet, row, header, alignments, baseStyle)
}
\arguments{
\item{sheet}{Sheet to write the table header to.}

\item{row}{Row of sheet to write the table header to.}

\item{header}{Characters vector containing table header columns.}

\item{alignments}{Vector of values from `{r, l, c}` for cell.}

\item{baseStyle}{Base style for the cells}
}
\value{
Next row number.
}
\description{
Writes the table header to the `row` of `sheet` with given
alignments.
}
