% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{xDecafPreemble}
\alias{xDecafPreemble}
\title{A function to provide the enriched trade data-frame from source session.}
\usage{
xDecafPreemble(
  sourceSession,
  targetSession,
  containerMap,
  containerType,
  gte = NULL
)
}
\arguments{
\item{sourceSession}{The rdecaf session for source instance.}

\item{targetSession}{The rdecaf session for target instance.}

\item{containerMap}{TODO:}

\item{containerType}{TODO:}

\item{gte}{The commitment date after which the trades should be considered. Default is NULL.}
}
\value{
A trade data-frame with mapped/created accmains and auxiliary resource information.
}
\description{
This is the description
}
