% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocketBase.R
\name{xtractDat}
\alias{xtractDat}
\title{Pocketbase dependent function to convert a list element to a data frame by ticker and exchange info.}
\usage{
xtractDat(list, element, dat)
}
\arguments{
\item{list}{the list containing the element.}

\item{element}{string of the name of the item to extract from the list, e.g. estimates.}

\item{dat}{the data frame to join with in order to get the primary key symbol.}
}
\value{
a data frame of the stacked records.
}
\description{
This is the description
}
