% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{accountPreembleMethod2}
\alias{accountPreembleMethod2}
\title{This function applies the account preemble method 1.}
\usage{
accountPreembleMethod2(accounts, sysAccs, custodian, session)
}
\arguments{
\item{sysAccs}{A data frame with the system accounts.}

\item{custodian}{The custodian id.}

\item{session}{The rdecaf session.}

\item{records}{A data frame with columns ACCNAME, targetPortName and REFCCY.}
}
\value{
Returns the side by side comparison of decaf position and provider position.
}
\description{
Accounts are created with the name as provided if such
name is missing.If target portfolio name does not exist
yet, function creates. If such portfolio assignment is missing,
this functions creates a portfolio with the same name
as account.
}
