% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{alertEmail}
\alias{alertEmail}
\title{A function that sends out a generic DECAF email alert given parameters.}
\usage{
alertEmail(
  session,
  data = NULL,
  emailParams,
  dtPath = "details.csv",
  snd = Sys.Date(),
  wdw = c("00:00:00", "23:59:00"),
  dte = NULL,
  tz = "Asia/Singapore",
  noBody = FALSE,
  ...
)
}
\arguments{
\item{session}{list of rdecaf session.}

\item{data}{list containing two optional data frame elements (summary or first element displays in body, details or second in attachments). Default NULL for just a message.}

\item{emailParams}{list of the parameters for the email dispatch. INCLUDE SUBJECT HERE.}

\item{dtPath}{the path containing the details/attachment data. Defaults to details.csv in the wd.}

\item{wdw}{vector of the time period in which to run the function. Defaults to 24H given a vector of c(gte,lte).}

\item{dte}{string of the weekday name of the day in which to run this alert. Defaults to NULL for all days.}

\item{tz}{string of the time-zone for gte and lte. Defaults to Asia/Singapore.}

\item{...}{Any additional arguments.}
}
\value{
Sends email with the alert.
}
\description{
This is the description
}
\examples{
df <- data.frame(parent=c(rep("A",30),rep("B",30),rep("c",30)),child=sample(1:10000,90))
emailParams <- list("sendEmail" = TRUE,"from" = "info@telostat.com","emailList" = "andre@telostat.com","isLocal" = FALSE, "subject" = "Example Alert")
alertEmail(session=session,data=list("summary"=df \%>\% dplyr::group_by(parent) \%>\% dplyr::summarize(v=sum(child)),"details"=df),emailParams=emailParams,wdw=c("23:30:00","23:59:00"),dte="Saturday")
}
