% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnl.R
\name{bare_get}
\alias{bare_get}
\title{Gets the pnl endpoint from decaf.}
\usage{
bare_get(
  apiURL = "/apis/function/valuation-reports/eventsreport",
  params = list(),
  session = NULL
)
}
\arguments{
\item{apiURL}{the npl endpoint API URL string.}

\item{params}{a list of additional parameters.}

\item{session}{the decaf session.}
}
\value{
A list with the pnl ledger data.
}
\description{
This is the description
}
