% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{beanbagNAVTable}
\alias{beanbagNAVTable}
\title{Prepares the NAV data frame.}
\usage{
beanbagNAVTable(x, inception, pxinfo)
}
\arguments{
\item{x}{A list with "portfolio", "ccy", "currentNAV", "previous NAV".}

\item{inception}{The launch date.}

\item{pxinfo}{The YTD Performance.}
}
\value{
A data frame.
}
\description{
This is a description.
}
