% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{decafPushActions}
\alias{decafPushActions}
\title{Syncs actions in batches.}
\usage{
decafPushActions(actions, session, N = 500, omitFlag = NULL)
}
\arguments{
\item{actions}{The actions data frame.}

\item{session}{The rdecaf session.}

\item{N}{The batch size.}

\item{omitFlag}{The flag number to be ignored. Default NULL}
}
\description{
This is the description
}
