% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{decafSyncResources}
\alias{decafSyncResources}
\title{This function syncs the resources between source DECAF and target DECAF.}
\usage{
decafSyncResources(
  tSession,
  sSession,
  sAccountNames,
  overrideTypes = NULL,
  customSymbolFunction = function(x) {     paste(x[, "isin"], x[, "ccymain"], "[CUB]")
    },
  tResources = NULL,
  matchBy = "symbol",
  override = FALSE
)
}
\arguments{
\item{tSession}{The target session.}

\item{sSession}{The source session.}

\item{sAccountNames}{The account names at source.}

\item{overrideTypes}{A vector with strings defining the resource ctypes to be overriden. Default NULL}

\item{customSymbolFunction}{Either a function or NULL.}

\item{tResources}{The resource data frame at target.}

\item{matchBy}{The field to match by.}

\item{override}{Shall the resources be overidden? Default FALSE.}
}
\value{
The target resources data frame.
}
\description{
This is the description
}
