% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{decafSyncTrades}
\alias{decafSyncTrades}
\title{This function syncs the trades between source DECAF and target DECAF.}
\usage{
decafSyncTrades(
  accounts,
  sSession,
  tSession,
  resources,
  gte,
  omitFlag = NULL,
  sResources = NULL,
  customSymbolFunction = NULL
)
}
\arguments{
\item{accounts}{The accounts list.}

\item{sSession}{The target session.}

\item{tSession}{The source session.}

\item{resources}{The resources data frame at target.}

\item{gte}{The greater than or equal to date for commitment.}

\item{omitFlag}{The flag type of record to bypass sync. Default NULL.}

\item{sResources}{The resources data frame at source.}

\item{customSymbolFunction}{Either NULL or a function. Default NULL>}
}
\value{
NULL.
}
\description{
This is the description
}
