% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{emailHTMLTable}
\alias{emailHTMLTable}
\title{A function to generate email-friendly HTML table.}
\usage{
emailHTMLTable(
  df,
  provider,
  caption,
  sourceType = "API",
  rowGroups = NULL,
  rowBGColor = list(),
  collapse = "separate",
  spacing = "4px",
  footer = NULL
)
}
\arguments{
\item{df}{A data-frame.}

\item{provider}{The name of the provider of the data to be used in the footer.}

\item{caption}{The caption of the table to be used.}

\item{sourceType}{The type of the source data to be used for the footer.E.g., 'Api'.}

\item{rowGroups}{A list of two elements of the same length. The first
element will contain the row spans (i.e. a numeric atomic vector) whereas
the second element will contain the names (i.e. a character atomic vector).
See the examples for more info. Defauls to NULL}

\item{rowBGColor}{A list of two elements with the row indices for alternative background color.
Default in NULL in which case it constructs itself.}

\item{collapse}{The table collapse parameters. Default is separate.}

\item{spacing}{The cell spacing of the table. Default is 4px.",}

\item{footer}{The footer to be used. Default is NULL in which case it constructs itself.}
}
\value{
A java-object of class org.apache.commons.mail.SimpleEmail
}
\description{
TODO:
}
