% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{executeValuationAndPool}
\alias{executeValuationAndPool}
\title{This function executes a job give a portfolio id and pools the result until finished.}
\usage{
executeValuationAndPool(id, session, maxWait = 900)
}
\arguments{
\item{id}{The portfolio id.}

\item{session}{The rdecaf session.}

\item{maxWait}{The maximum waiting time in seconds.}
}
\description{
This is the description
}
