% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extendFileName}
\alias{extendFileName}
\title{A function to extend file name by random/date string.}
\usage{
extendFileName(str, customExt = NULL, randomStrN = NULL,
  addTime = TRUE)
}
\arguments{
\item{str}{A string.}

\item{customExt}{A custom extention of the string. Default is NULL.}

\item{randomStrN}{The random string character size. Default is NULL.}

\item{addTime}{Boolean to indicate whether time should be added. Default is TRUE.}
}
\value{
The extended file name.
}
\description{
A function to extend file name by random/date string.
}
