% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{getAccountWiseTrades}
\alias{getAccountWiseTrades}
\title{A function to get account-wise trades from a DECAF instance.}
\usage{
getAccountWiseTrades(accounts, session, gte = NULL, dateType = "commitment")
}
\arguments{
\item{accounts}{A vector with account id's.}

\item{session}{The DECAF session info.}

\item{gte}{The date after which trades should be considered.}

\item{dateType}{The type, either commitment, settlement, created or updated. Default is commitment.}
}
\value{
A data-frame with DECAF trades.
}
\description{
This is the description
}
