% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{getAssetEvolution}
\alias{getAssetEvolution}
\title{Gets the asset evolution for a portfolio from pconsolidation}
\usage{
getAssetEvolution(portfolio, date, years = "2", session, ...)
}
\arguments{
\item{portfolio}{The portfolio id.}

\item{date}{The date__lte.}

\item{years}{The number of years to look back.}

\item{session}{The rdecaf session,}

\item{...}{Any additional parameters.}
}
\value{
An data frame with NAV and AUM.
}
\description{
This is a description.
}
