% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{getAssetReturns}
\alias{getAssetReturns}
\title{Provides the asset returns MTD and YTD including FX returns}
\usage{
getAssetReturns(portfolio, date, ccy, resources, session)
}
\arguments{
\item{portfolio}{The portfolio id.}

\item{date}{The date of consideration.}

\item{ccy}{The portfolio currency.}

\item{resources}{The resource data frame.}

\item{session}{The rdecaf session.}
}
\value{
A data frame with the asset returns.
}
\description{
This is the description
}
