% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{getAssetReturns}
\alias{getAssetReturns}
\title{Provides the asset returns MTD and YTD including FX returns}
\usage{
getAssetReturns(
  date,
  ccy,
  resources,
  priceField = "ohlcID",
  periods,
  returnOnly,
  excludeWeekends,
  session,
  treat = TRUE,
  exclude = c("FXFWD", "DEPO", "LOAN")
)
}
\arguments{
\item{date}{The date of consideration.}

\item{ccy}{The portfolio currency.}

\item{resources}{The resource data frame.}

\item{priceField}{The name of the price field in resources.}

\item{periods}{A vector with periods: c('DTD', 'WTD', 'MTD', 'QTD', 'YTD').}

\item{returnOnly}{Consider only Total Return?}

\item{excludeWeekends}{Should the weekends be excluded? Default to TRUE.}

\item{session}{The rdecaf session.}

\item{treat}{Should the time series be treated? Default=TRUE.}

\item{exclude}{Which ctypes should be excluded? Default=c("FXFWD", "DEPO", "LOAN").}
}
\value{
A data frame with the asset returns.
}
\description{
This is the description
}
