% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdings.R
\name{getConsolidationFromContainerName}
\alias{getConsolidationFromContainerName}
\title{Gets the consolidation based on container type and name}
\usage{
getConsolidationFromContainerName(containerNames, containerType, ccy, date,
  session)
}
\arguments{
\item{containerNames}{A vector with container names}

\item{containerType}{The container type}

\item{ccy}{The currency of the consolidation}

\item{date}{The date of the consolidation}

\item{session}{The rdecaf session}
}
\value{
A list with the holdings data-frames.
}
\description{
This is the description
}
