% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{getDBObject}
\alias{getDBObject}
\title{This function gets the data base object as data frame.}
\usage{
getDBObject(endpoint, session, addParams = NULL)
}
\arguments{
\item{endpoint}{The endpoint.}

\item{session}{The rdecaf session.}

\item{addParams}{A list with additional parameters.}
}
\value{
A data frame with the endpoint as data frame.
}
\description{
This is the description
}
