% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{getHoldingsDetails}
\alias{getHoldingsDetails}
\title{Prepares the holdings detail data frame.}
\usage{
getHoldingsDetails(holdings, colSelect, nav = NULL)
}
\arguments{
\item{holdings}{The holdings data frame.}

\item{colSelect}{Which columns to select.}

\item{nav}{The net asset value.}
}
\value{
A list with holdings details information.
}
\description{
This is a description.
}
