% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{getHoldingsSummary}
\alias{getHoldingsSummary}
\title{Prepares the holdings summary data frame.}
\usage{
getHoldingsSummary(holdings)
}
\arguments{
\item{holdings}{The holdings data frame}
}
\value{
A list with holdings summary information.
}
\description{
This is a description.
}
