% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{getInitialPremium}
\alias{getInitialPremium}
\title{A function to get initial premium date and value for a trades data frame.}
\usage{
getInitialPremium(data)
}
\arguments{
\item{data}{a trades data frame}
}
\value{
A list
}
\description{
This is the description
}
