% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdings.R
\name{getNestedHoldings}
\alias{getNestedHoldings}
\title{A function to generate a nested holdings data-frame}
\usage{
getNestedHoldings(
  holdings,
  levels,
  toplevel = "Subtype",
  sublevels = c("CCY", "Country"),
  orderByKeys = c("Expiry", "CCY", "Exposure", "Name")
)
}
\arguments{
\item{holdings}{The ordered holdings.}

\item{levels}{TODO}

\item{toplevel}{TODO}

\item{sublevels}{TODO}

\item{orderByKeys}{Based on which columns should the final grouping be ordered?Defaults: Expiry, CCY, Exposure, Name and decreasing TRUE?}
}
\value{
Returns the nested holdings data-frame.
}
\description{
This is the description
}
