% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{getPerformance}
\alias{getPerformance}
\title{Gets and prepares the performance information for a portfolio.}
\usage{
getPerformance(
  portfolio,
  start,
  end,
  freq = "daily",
  session,
  period = "Y-0",
  benchMark = NULL,
  rF = NULL
)
}
\arguments{
\item{portfolio}{The portfolio id.}

\item{start}{The desired start date.}

\item{end}{The desired end date.}

\item{freq}{The desired frequency.}

\item{session}{The rdecaf session.}

\item{period}{The period memnonic. Default is Y-0.}

\item{benchMark}{a list containing the elements from getBenchmark above. Defaults to NULL.}

\item{rF}{the risk free data. Defaults to NULL.}
}
\value{
A list with the indexed cumulative return series and performance stats.
}
\description{
This is a description.
}
