% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanbag.R
\name{getPreviousNAV}
\alias{getPreviousNAV}
\title{Gets the previous NAN either using external or internal valuations.}
\usage{
getPreviousNAV(portfolio, date, ccy, period, external, session,
  inception = as.Date("1900-01-01"))
}
\arguments{
\item{portfolio}{The portfolio id.}

\item{date}{The report date.}

\item{ccy}{The portoflio currency.}

\item{period}{The period memnonic, i.e Y-0 for last year end.}

\item{external}{TRUE if external valuation is to be used.}

\item{session}{The rdecaf session.}

\item{inception}{The inception date. Default is 1900-01-01}
}
\value{
A list.
}
\description{
This is a description.
}
