% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{getSlicedOhlcs}
\alias{getSlicedOhlcs}
\title{A function to retrieve and slice OHLC observations as per given period.}
\usage{
getSlicedOhlcs(symbols, session, date, periods, excludeWeekends = TRUE)
}
\arguments{
\item{symbols}{A vector of ohlcs symbols.}

\item{session}{The rdecaf session.}

\item{date}{The report date.}

\item{periods}{A vector with desired period slices: 'DTD', 'WTD', 'MTD', 'QTD', 'YTD'}

\item{excludeWeekends}{Should the weekends be excluded? Default to TRUE.}
}
\value{
A list with the slices per period.
}
\description{
This is the description
}
