% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnl.R
\name{getStocksAndAssets}
\alias{getStocksAndAssets}
\title{flattens the pnl endpoint list data from decaf.}
\usage{
getStocksAndAssets(session)
}
\arguments{
\item{session}{the decaf session.}
}
\value{
A flat df containing the session artifacts and associated asset class.
}
\description{
This is the description
}
