% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{getTransferValueAmounts}
\alias{getTransferValueAmounts}
\title{A function to get the value amounts of transfers for a portfolio.}
\usage{
getTransferValueAmounts(portfolio, since, until, currency, session)
}
\arguments{
\item{portfolio}{The portfolio id.}

\item{since}{The date since.}

\item{until}{The date until.}

\item{currency}{The portfolio currency.}

\item{session}{The rdecaf session.}
}
\value{
A data frame with the converted value amounts of transfers.
}
\description{
This is the description
}
