% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mgrep}
\alias{mgrep}
\title{A function to grep multiple keys in vector.}
\usage{
mgrep(x, keys, useMatch = FALSE)
}
\arguments{
\item{x}{A vector with strings.}

\item{keys}{A vector with keys to grep in x.}

\item{useMatch}{Shall match be used instead of grep?}
}
\value{
A 'x' by 'keys' size data frame with the grepped keys.
}
\description{
This is a description.
}
