% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health.R
\name{ohlcHealth}
\alias{ohlcHealth}
\title{This function checks the health of the OHLC observations.}
\usage{
ohlcHealth(ohlc, asof = Sys.Date())
}
\arguments{
\item{ohlc}{A data-frame with the ohlc observations.}

\item{asof}{The date to compare ohlc observations to.}
}
\value{
A data-frame with state of ohlc series.
}
\description{
This is a description
}
