% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{percentify}
\alias{percentify}
\title{A function for formatting percentages:}
\usage{
percentify(number)
}
\arguments{
\item{number}{An array of numeric values.}
}
\value{
A string of percentage:
}
\description{
A function for formatting percentages:
}
