% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performancefee.R
\name{performancePreemble}
\alias{performancePreemble}
\title{A function to provide the performance fee preemble.}
\usage{
performancePreemble(perfMeta, session)
}
\arguments{
\item{perfMeta}{A list with the performance fee meta data.}

\item{session}{The rdecaf session.}
}
\value{
A xts data frame with the performance fee preemble.
}
\description{
This is the description
}
