% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{preparePortfolioPayload}
\alias{preparePortfolioPayload}
\title{This function prepares the portfolio payload as data frame.}
\usage{
preparePortfolioPayload(accounts, teams, rccy = "USD", session)
}
\arguments{
\item{accounts}{The accounts list.}

\item{teams}{The teams data frame.}

\item{rccy}{The fallback value for the portfolios currency.}

\item{session}{The session.}
}
\value{
The portfolio payload as data frame.
}
\description{
This is the description
}
