% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safeNull}
\alias{safeNull}
\title{A function to replace null and length 0 value with NA}
\usage{
safeNull(x, replace = NA)
}
\arguments{
\item{x}{A variable}

\item{replace}{The value to be returned if NULL. Default = NA.}
}
\value{
Returns x or replace.
}
\description{
This is the description
}
