% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbg.R
\name{treatBBGResultsPrice}
\alias{treatBBGResultsPrice}
\title{A function to treat the results of a bbgdl price request.}
\usage{
treatBBGResultsPrice(result, reqData, noCents = TRUE)
}
\arguments{
\item{result}{The result data frame as obtained from bdhls.}

\item{reqData}{The original request data from getRequestDataPrice.}

\item{noCents}{If noCents, prices in cents will be divided by 100. Default is TRUE.}
}
\value{
A list with the ohlc observations and invalid symbols.
}
\description{
A function to treat the results of a bbgdl price request.
}
