% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{xDecafPreemble}
\alias{xDecafPreemble}
\title{A function to provide the enriched trade data-frame from source session.}
\usage{
xDecafPreemble(sourceSession, targetSession, portfolioMap)
}
\arguments{
\item{sourceSession}{The rdecaf session for source instance.}

\item{targetSession}{The rdecaf session for target instance.}

\item{portfolioMap}{A named character vector, such as c("[NAME of source portfolio]"="[NAME of target portfolio]")}
}
\value{
A trade data-frame with mapped/created accmains and auxiliary resource information.
}
\description{
This is the description
}
